refactor(server): split http.rs into feature modules
Replace the 3100-line http.rs monolith with a feature-oriented http/
package (rust-skills proj-mod-by-feature): state, error, export, health,
transcribe, stream, openai_api, admin, jobs_api, plus extracted unit
tests. Public paths (http::transcribe, http::AppState, …) stay stable
for the router and jobs/ws callers; behavior is unchanged.
