--- crates/gigastt/src/server/http.rs
@@ -1,3105 +1,0 @@
-//! HTTP handlers for REST API endpoints.
-
-use axum::body::Bytes;
-use axum::extract::{Multipart, Query, State};
-use axum::http::StatusCode;
-use axum::http::header;
-use axum::response::sse::{Event, KeepAlive, Sse};
-use axum::response::{IntoResponse, Json, Response};
-use futures_util::StreamExt;
-use futures_util::stream::Stream;
-use serde::{Deserialize, Serialize};
-use std::str::FromStr;
-use std::sync::Arc;
-
-use arc_swap::ArcSwap;
-
-use super::config::{RuntimeLimits, pool_retry_after_ms, pool_retry_after_secs};
-use super::jobs::{JobEvent, JobQueue, JobStatus, JobStore};
-use super::metrics::MetricsRegistry;
-use gigastt_core::export::{ExportFormat, RenderOpts, Segment};
-use gigastt_core::inference::Engine;
-
-/// Shared application state for all handlers. Carries runtime limits so the
-/// WebSocket path can enforce configurable frame / idle bounds without
-/// re-threading every CLI arg through each handler, plus an optional
-/// in-tree `MetricsRegistry` backing the `/metrics` endpoint.
-///
-/// Also carries a shutdown `CancellationToken` and a `TaskTracker` used to
-/// drain in-flight WebSocket / SSE tasks on SIGTERM. `axum::serve`'s
-/// built-in `with_graceful_shutdown` only tracks the HTTP router; upgraded
-/// WebSocket handlers and `spawn_blocking` SSE tasks fall outside that lane
-/// and must be drained explicitly.
-pub struct AppState {
-    /// The live inference engine, held behind an [`ArcSwap`] so it can be
-    /// atomically replaced by the model hot-reload endpoint without a restart.
-    /// Handlers `load_full()` the current `Arc<Engine>` at request start and
-    /// use it for the whole request; a concurrent swap only affects requests
-    /// that start after it, so in-flight work always finishes against the
-    /// engine (and pool) it began with.
-    pub engine: Arc<ArcSwap<Engine>>,
-    /// Rebuilds the engine from the exact boot recipe (model dir, pool sizes,
-    /// threads, punctuation / ITN / VAD / hotwords). `Some` on the real server
-    /// path; `None` for the thin `run`/`run_with_shutdown` test entry points
-    /// and unit tests, where the reload endpoint reports `reload_unsupported`.
-    pub engine_builder: Option<EngineBuilder>,
-    /// Serializes model reloads so two concurrent `POST /v1/admin/reload`
-    /// calls can't both rebuild + swap; the loser gets `409 reload_in_progress`.
-    pub reload_lock: Arc<tokio::sync::Mutex<()>>,
-    pub limits: Arc<ArcSwap<RuntimeLimits>>,
-    pub metrics_registry: Option<Arc<MetricsRegistry>>,
-    pub shutdown: tokio_util::sync::CancellationToken,
-    pub tracker: tokio_util::task::TaskTracker,
-    /// In-memory job store and queue. `Some` only when `--enable-jobs` is set;
-    /// handlers for `/v1/jobs` are registered conditionally and expect this to
-    /// be populated, but the shared state is `Option` so non-job builds compile.
-    pub jobs: Option<JobServerState>,
-}
-
-/// State shared by the `/v1/jobs` handlers. Kept behind `Arc` so the executor,
-/// queue, and handlers all reference the same store.
-#[derive(Clone)]
-pub struct JobServerState {
-    pub store: Arc<dyn JobStore>,
-    pub queue: Arc<JobQueue>,
-}
-
-/// Recipe that rebuilds a fully-configured [`Engine`] from the operator's boot
-/// options. Stored in [`AppState`] so `POST /v1/admin/reload` can produce a
-/// fresh engine that re-applies the punctuation / ITN / VAD / hotword chain —
-/// a bare `Engine::load_*` starts with all of those set to `None`.
-pub type EngineBuilder = Arc<dyn Fn() -> anyhow::Result<Engine> + Send + Sync>;
-
-/// GET /metrics — Prometheus text-format exposition. Returns 404 when the
-/// server was started without `--metrics`.
-pub async fn metrics(State(state): State<Arc<AppState>>) -> Response {
-    match &state.metrics_registry {
-        Some(registry) => (
-            StatusCode::OK,
-            [(
-                header::CONTENT_TYPE,
-                "text/plain; version=0.0.4; charset=utf-8",
-            )],
-            registry.render_prometheus(),
-        )
-            .into_response(),
-        None => (
-            StatusCode::NOT_FOUND,
-            Json(serde_json::json!({
-                "error": "metrics endpoint disabled",
-                "code": "metrics_disabled",
-            })),
-        )
-            .into_response(),
-    }
-}
-
-/// Health check response.
-#[derive(Serialize)]
-pub struct HealthResponse {
-    /// Always `"ok"` when the server is running.
-    pub status: String,
-    /// Stable model identifier for the head actually loaded
-    /// (`"gigaam-v3-rnnt"` or `"gigaam-v3-e2e-rnnt"`).
-    pub model: String,
-    /// Recognition head in use: `"rnnt"` or `"e2e_rnnt"`. Added so a client can
-    /// tell from a single `/health` call which head (and therefore which output
-    /// style) is producing transcripts.
-    pub variant: String,
-    /// Server version from `CARGO_PKG_VERSION`.
-    pub version: String,
-    /// Whether the punctuation / casing restoration pass is active for this
-    /// server (the effective `--punctuation` policy).
-    pub punctuation: bool,
-    /// Whether inverse text normalization (numbers → digits) is active for this
-    /// server (the effective `--itn` policy).
-    pub itn: bool,
-}
-
-/// Model info response.
-#[derive(Serialize)]
-pub struct ModelInfo {
-    /// Stable model identifier for the head actually loaded
-    /// (`"gigaam-v3-rnnt"` or `"gigaam-v3-e2e-rnnt"`).
-    pub id: String,
-    /// Human-readable model name.
-    pub name: String,
-    /// Recognition head in use: `"rnnt"` or `"e2e_rnnt"`.
-    pub variant: String,
-    /// Server version from `CARGO_PKG_VERSION`.
-    pub version: String,
-    /// Encoder precision in use: `"int8"` or `"fp32"`.
-    pub encoder: String,
-    /// Number of tokens in the BPE vocabulary.
-    pub vocab_size: usize,
-    /// Native sample rate the model operates at (always 16000 Hz).
-    pub sample_rate: u32,
-    /// Total number of session triplets in the pool.
-    pub pool_size: usize,
-    /// Number of session triplets currently available for checkout.
-    pub pool_available: usize,
-    /// Audio container formats accepted by `/v1/transcribe`.
-    pub supported_formats: Vec<String>,
-    /// PCM sample rates accepted by the WebSocket endpoint.
-    pub supported_rates: Vec<u32>,
-    /// Whether the punctuation / casing restoration pass is active (effective
-    /// `--punctuation` policy for the loaded head).
-    pub punctuation: bool,
-    /// Whether inverse text normalization (numbers → digits) is active
-    /// (effective `--itn` policy for the loaded head).
-    pub itn: bool,
-    /// Whether speaker diarization is available (feature-gated build + model loaded).
-    /// Added in v0.7.0 so clients can probe capabilities via REST instead of
-    /// opening a WebSocket just to read the `Ready` frame.
-    pub diarization: bool,
-}
-
-/// Transcription response.
-#[derive(Serialize)]
-pub struct TranscribeResponse {
-    /// Full recognized transcript text.
-    pub text: String,
-    /// Word-level timing, confidence, and optional speaker annotations.
-    pub words: Vec<gigastt_core::inference::WordInfo>,
-    /// Duration of the submitted audio in seconds.
-    pub duration: f64,
-    /// Mean confidence across all words (duration-weighted average of
-    /// `words[].confidence`): an average of per-word softmax scores, not a
-    /// calibrated transcript probability. Additive; omitted when no words
-    /// were decoded so the pre-field response shape is preserved.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub confidence: Option<f32>,
-    /// Transcript segments grouped from word timestamps, present only when the
-    /// caller passed `?segments=true`. Additive: absent from the default response,
-    /// so existing clients that read `text` / `words` / `duration` are unaffected.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub segments: Option<Vec<Segment>>,
-}
-
-/// Query parameters for `/v1/transcribe` export formatting.
-#[derive(Debug, Default, Clone, Deserialize)]
-pub struct ExportParams {
-    /// Export format: `json` (default), `txt`, `srt`, `vtt`, `md`.
-    pub format: Option<String>,
-    /// When set, the response carries `Content-Disposition: attachment` with this
-    /// filename (or `transcript.<ext>` if the value is empty).
-    pub download: Option<String>,
-    /// Maximum characters per subtitle/caption line. `0` = unlimited.
-    #[serde(default)]
-    pub max_chars_per_line: Option<usize>,
-    /// Maximum words per subtitle/caption line. `0` = unlimited.
-    #[serde(default)]
-    pub max_words_per_line: Option<usize>,
-    /// Include per-word timestamps in Markdown output.
-    #[serde(default)]
-    pub word_timestamps: Option<bool>,
-    /// Return cue-sized segments. In the default (JSON) response this adds a
-    /// `segments` array; combined with `format=md` it switches Markdown to
-    /// `### [mm:ss]` segment headers. Ignored for `txt`/`srt`/`vtt` (those are
-    /// already flat / cue-based).
-    #[serde(default)]
-    pub segments: Option<bool>,
-    /// Per-request override for the punctuation / casing restoration pass.
-    /// `Some(true)` forces it on (409 `punctuation_not_available` if no
-    /// punctuation model is loaded), `Some(false)` skips it, absent = the
-    /// server's boot default. Applies to `POST /v1/transcribe` only.
-    #[serde(default)]
-    pub punctuation: Option<bool>,
-    /// Per-request override for inverse text normalization (number-words →
-    /// digits). `Some(true)`/`Some(false)` force the state, absent = boot
-    /// default. Pure code (no model), so always accepted. `POST /v1/transcribe`
-    /// only.
-    #[serde(default)]
-    pub itn: Option<bool>,
-    /// Per-request override for VAD gating. `Some(true)` decodes only detected
-    /// speech (409 `vad_not_loaded` if no VAD is loaded), `Some(false)` decodes
-    /// the whole buffer, absent = boot default. `POST /v1/transcribe` only.
-    #[serde(default)]
-    pub vad: Option<bool>,
-    /// Per-request hotword phrases, comma-separated
-    /// (`?hotwords=сбер,тинькофф`). Absent = keep the engine boot biaser;
-    /// present with an empty value (or only empty segments) forces biasing
-    /// off for this request; non-empty replaces the biaser for this request
-    /// only. Capped at 64 phrases / 64 chars each (400 on overflow).
-    /// `POST /v1/transcribe` only.
-    #[serde(default)]
-    pub hotwords: Option<String>,
-    /// Additive logit boost for per-request hotwords. Only applied when
-    /// `hotwords` is also present; absent defaults to 5.0.
-    /// `POST /v1/transcribe` only.
-    #[serde(default)]
-    pub hotwords_boost: Option<f32>,
-    /// Forward-compatibility guard for a future multi-model server: names the
-    /// recognition head the request expects. A single-variant engine can't
-    /// switch, so any value other than the loaded variant returns 409
-    /// `variant_not_loaded`; matching (or absent) proceeds. `POST /v1/transcribe`
-    /// only.
-    #[serde(default)]
-    pub variant: Option<String>,
-    /// Channel handling for file transcription. `split` transcribes left/right
-    /// channels as separate speakers. Defaults to mono mix.
-    #[serde(default)]
-    pub channels: Option<String>,
-    /// Request speaker diarization. Mutually exclusive with `channels=split`.
-    #[serde(default)]
-    pub diarization: Option<bool>,
-    /// Raw headerless telephony codec of the upload body: `pcmu` (alias
-    /// `ulaw`), `pcma` (alias `alaw`), or `g722`. When set, the body is
-    /// decoded as a raw byte stream of that codec instead of sniffing a
-    /// container — for RTP dumps and Asterisk Monitor raw captures. Requires
-    /// `sample_rate`.
-    #[serde(default)]
-    pub codec: Option<String>,
-    /// Sample rate (Hz) of a raw `codec` upload — mandatory when `codec` is
-    /// set (typical telephony: 8000). G.722 decodes to its native 16 kHz
-    /// regardless; both 8000 (the SDP clock-rate convention) and 16000 are
-    /// accepted for it.
-    #[serde(default)]
-    pub sample_rate: Option<u32>,
-}
-
-/// Split a comma-separated `?hotwords=` value into trimmed non-empty phrases.
-/// Empty input (or only empty segments) yields an empty list, which the engine
-/// treats as force-off for this request.
-pub(crate) fn parse_hotwords_query(raw: &str) -> Vec<String> {
-    raw.split(',')
-        .map(str::trim)
-        .filter(|p| !p.is_empty())
-        .map(str::to_string)
-        .collect()
-}
-
-/// Build [`TranscribeOverrides`] from REST query params. Absent knob params
-/// leave the corresponding override field as `None` (engine boot default).
-pub(crate) fn overrides_from_export_params(
-    params: &ExportParams,
-) -> gigastt_core::inference::TranscribeOverrides {
-    gigastt_core::inference::TranscribeOverrides {
-        punctuation: params.punctuation,
-        itn: params.itn,
-        vad: params.vad,
-    }
-}
-
-/// Build optional [`HotwordOverride`] from REST query params. `None` when the
-/// `hotwords` key is absent (engine boot biaser). `Some` when the key is
-/// present — even empty — so `?hotwords=` can force biasing off.
-pub(crate) fn hotwords_from_export_params(
-    params: &ExportParams,
-) -> Option<gigastt_core::inference::HotwordOverride> {
-    params.hotwords.as_ref().map(|raw| {
-        gigastt_core::inference::HotwordOverride::new(
-            parse_hotwords_query(raw),
-            // Boost only applies when hotwords is present; ignore alone.
-            params.hotwords_boost,
-        )
-    })
-}
-
-/// Render a transcription result into the requested export format.
-///
-/// Returns `None` when the caller explicitly requested the default JSON
-/// response, so the handler can keep serving the existing `TranscribeResponse`
-/// contract unchanged.
-#[allow(clippy::result_large_err)]
-fn render_export_response(
-    result: &gigastt_core::inference::TranscribeResult,
-    params: &ExportParams,
-) -> Result<Option<Response>, ApiError> {
-    let format_str = params.format.as_deref().unwrap_or("json");
-    if format_str.eq_ignore_ascii_case("json") {
-        return Ok(None);
-    }
-
-    let format = ExportFormat::from_str(format_str)
-        .map_err(|e| api_error(StatusCode::BAD_REQUEST, &format!("{e}"), "invalid_format"))?;
-
-    let opts = RenderOpts {
-        max_chars_per_line: params.max_chars_per_line.unwrap_or(80),
-        max_words_per_line: params.max_words_per_line.unwrap_or(14),
-        include_word_timestamps: params.word_timestamps.unwrap_or(false),
-    };
-
-    // Precedence for `format` × `segments`: only Markdown composes with
-    // `segments=true`, switching to `### [mm:ss]` section headers over the same
-    // cue boundaries as SRT/VTT. `txt`/`srt`/`vtt` ignore `segments` (flat /
-    // already cue-based); plain `format=md` keeps the frontmatter + `# Transcript`
-    // blob unchanged.
-    let body = if format == ExportFormat::Md && params.segments.unwrap_or(false) {
-        gigastt_core::export::to_md_segments(
-            result,
-            opts.max_chars_per_line,
-            opts.max_words_per_line,
-        )
-    } else {
-        format.render(result, &opts)
-    };
-    let mut builder = Response::builder()
-        .status(StatusCode::OK)
-        .header(header::CONTENT_TYPE, format.content_type());
-
-    if let Some(filename) = &params.download {
-        let filename = if filename.is_empty() {
-            format!("transcript.{}", format.extension())
-        } else {
-            filename.clone()
-        };
-        // The filename is user-controlled (query param), so emit it as an
-        // RFC 6266 value: quotes/semicolons can no longer inject extra header
-        // parameters (filename spoofing) and non-ASCII names survive via
-        // `filename*`. The helper output is pure ASCII, which makes the header
-        // conversion infallible; keep the defensive fallback anyway.
-        let disposition = header::HeaderValue::from_str(&content_disposition_attachment(&filename))
-            .unwrap_or_else(|_| {
-                header::HeaderValue::from_str(&format!(
-                    "attachment; filename=\"transcript.{}\"",
-                    format.extension()
-                ))
-                .expect("static content-disposition is always a valid header value")
-            });
-        builder = builder.header(header::CONTENT_DISPOSITION, disposition);
-    }
-
-    let response = builder.body(axum::body::Body::from(body)).map_err(|e| {
-        api_error(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            &format!("failed to build response: {e}"),
-            "internal_error",
-        )
-    })?;
-    Ok(Some(response))
-}
-
-/// Build a `Content-Disposition: attachment` value for a user-controlled
-/// filename per RFC 6266. The legacy quoted `filename=` parameter carries an
-/// ASCII fallback (`"` / `\` / control / non-ASCII characters replaced by
-/// `_`), and the full name travels percent-encoded in `filename*=UTF-8''…`
-/// (RFC 5987 attr-char set emitted verbatim, everything else as `%XX` UTF-8
-/// bytes). Without the fallback sanitization a name like `x"; filename*=…`
-/// would splice extra parameters into the header — `HeaderValue` only rejects
-/// control characters, not quotes or semicolons — letting an attacker spoof
-/// the download filename seen by the client. The output is always printable
-/// ASCII, so `HeaderValue::from_str` accepts it unconditionally.
-fn content_disposition_attachment(filename: &str) -> String {
-    let mut fallback = String::with_capacity(filename.len());
-    for c in filename.chars() {
-        match c {
-            '"' | '\\' => fallback.push('_'),
-            c if c.is_ascii() && !c.is_ascii_control() => fallback.push(c),
-            _ => fallback.push('_'),
-        }
-    }
-    let mut encoded = String::with_capacity(filename.len());
-    for &b in filename.as_bytes() {
-        if b.is_ascii_alphanumeric() || b"!#$&+-.^_`|~".contains(&b) {
-            encoded.push(char::from(b));
-        } else {
-            encoded.push_str(&format!("%{b:02X}"));
-        }
-    }
-    format!("attachment; filename=\"{fallback}\"; filename*=UTF-8''{encoded}")
-}
-
-/// Error response produced by the REST handlers. Using `Response` directly
-/// (rather than a `(StatusCode, Json<_>)` tuple) lets timeout paths attach
-/// a `Retry-After` header without changing the handler signatures.
-type ApiError = Response;
-
-fn api_error(status: StatusCode, msg: &str, code: &str) -> ApiError {
-    (
-        status,
-        Json(serde_json::json!({"error": msg, "code": code})),
-    )
-        .into_response()
-}
-
-/// 503 response for pool-saturation backpressure: carries both the standard
-/// `Retry-After` header (seconds, per RFC 9110 §10.2.3) and a machine-readable
-/// `retry_after_ms` field in the JSON body so clients on either surface can
-/// back off with the same hint.
-fn api_timeout_error(limits: &RuntimeLimits) -> ApiError {
-    (
-        StatusCode::SERVICE_UNAVAILABLE,
-        [(
-            header::RETRY_AFTER,
-            pool_retry_after_secs(limits).to_string(),
-        )],
-        Json(serde_json::json!({
-            "error": "Server busy, try again later",
-            "code": "timeout",
-            "retry_after_ms": pool_retry_after_ms(limits),
-        })),
-    )
-        .into_response()
-}
-
-/// 503 response for the case where the pool was closed (graceful shutdown
-/// in progress). Distinct from `timeout` so clients can decide whether to
-/// retry: a closed pool is not coming back, so no `retry_after_ms` hint.
-fn api_pool_closed_error() -> ApiError {
-    (
-        StatusCode::SERVICE_UNAVAILABLE,
-        Json(serde_json::json!({
-            "error": "Server is shutting down",
-            "code": "pool_closed",
-        })),
-    )
-        .into_response()
-}
-
-/// 504 response for a single inference run that exceeded the per-request
-/// inference timeout (`--inference-timeout-secs`). Distinct from the pool
-/// `timeout` (503): the slot was free, the *run* itself was too slow / wedged,
-/// so there is no `Retry-After` — retrying the same payload would time out
-/// again. Extracted (mirroring [`api_timeout_error`]) so the status + code can
-/// be asserted without a model.
-fn api_inference_timeout_error() -> ApiError {
-    api_error(
-        StatusCode::GATEWAY_TIMEOUT,
-        "Inference timed out.",
-        "inference_timeout",
-    )
-}
-
-/// Resolve the `?codec=` / `?sample_rate=` query pair into a raw-decode recipe.
-/// `Ok(None)` means no `codec` was given and the body is probed as a container
-/// as before. Validation is done up front so a malformed request fails with a
-/// 400 before a pool slot is reserved. Kept separate from the handler so the
-/// status codes are unit-testable without a model.
-#[allow(clippy::result_large_err)]
-fn resolve_raw_codec(
-    params: &ExportParams,
-) -> Result<Option<(gigastt_core::inference::audio::TelephonyCodec, u32)>, ApiError> {
-    let Some(name) = params.codec.as_deref() else {
-        return Ok(None);
-    };
-    let codec =
-        gigastt_core::inference::audio::TelephonyCodec::from_name(name).ok_or_else(|| {
-            api_error(
-                StatusCode::BAD_REQUEST,
-                "Unsupported codec. Supported: pcmu (ulaw), pcma (alaw), g722",
-                "unsupported_codec",
-            )
-        })?;
-    let sample_rate = params.sample_rate.ok_or_else(|| {
-        api_error(
-            StatusCode::BAD_REQUEST,
-            "sample_rate is required when codec is set",
-            "invalid_sample_rate",
-        )
-    })?;
-    codec
-        .validate_sample_rate(sample_rate)
-        .map_err(|reason| api_error(StatusCode::BAD_REQUEST, &reason, "invalid_sample_rate"))?;
-    Ok(Some((codec, sample_rate)))
-}
-
-/// Decode a raw telephony byte stream and re-wrap it as an in-memory PCM16
-/// WAV, so every downstream engine path (mono, `channels=split`, diarized)
-/// works unchanged. Runs inside the blocking inference closure; errors map to
-/// the same 422 as container decode failures.
-fn raw_codec_to_wav(
-    body: &[u8],
-    codec: gigastt_core::inference::audio::TelephonyCodec,
-    sample_rate: u32,
-) -> Result<Bytes, gigastt_core::error::GigasttError> {
-    let samples = gigastt_core::inference::audio::decode_telephony_raw(body, codec, sample_rate)
-        .map_err(|e| gigastt_core::error::GigasttError::InvalidAudio {
-            reason: format!("{e:#}"),
-        })?;
-    Ok(Bytes::from(
-        gigastt_core::inference::audio::encode_wav_pcm16(&samples, 16000),
-    ))
-}
-
-/// Check out a session triplet from the engine's batch pool with the configured
-/// timeout and record the pool metrics, returning an owned reservation whose
-/// lifetime is detached (`into_owned`) so it can travel through `spawn_blocking`.
-///
-/// Both `/v1/transcribe` and `/v1/transcribe/stream` reserve a slot *before*
-/// decoding the upload; sharing the acquisition here keeps that backpressure —
-/// and the "reserve before decode" ordering that caps concurrent decodes at the
-/// pool size — identical across the two untrusted entry points. A saturated pool
-/// yields a 503 (`timeout`, with `Retry-After`) or `pool_closed`; the caller
-/// propagates it with `?`.
-async fn reserve_batch_slot(
-    engine: &Engine,
-    limits: &RuntimeLimits,
-    metrics: Option<&Arc<MetricsRegistry>>,
-) -> Result<
-    gigastt_core::inference::OwnedReservation<gigastt_core::inference::SessionTriplet>,
-    ApiError,
-> {
-    let checkout_start = std::time::Instant::now();
-    let guard = match tokio::time::timeout(
-        std::time::Duration::from_secs(limits.pool_checkout_timeout_secs),
-        engine.pool_for_batch().checkout(),
-    )
-    .await
-    {
-        Ok(Ok(guard)) => guard,
-        Ok(Err(_pool_closed)) => return Err(api_pool_closed_error()),
-        Err(_timeout) => {
-            if let Some(reg) = metrics {
-                reg.counter_inc("gigastt_pool_timeouts_total", &[], 1);
-                reg.histogram_record(
-                    "gigastt_pool_checkout_duration_seconds",
-                    &[],
-                    checkout_start.elapsed().as_secs_f64(),
-                );
-            }
-            return Err(api_timeout_error(limits));
-        }
-    };
-    if let Some(reg) = metrics {
-        reg.histogram_record(
-            "gigastt_pool_checkout_duration_seconds",
-            &[],
-            checkout_start.elapsed().as_secs_f64(),
-        );
-    }
-    Ok(guard.into_owned())
-}
-
-/// Per-segment error carried over the SSE channel: a stable machine-readable
-/// code plus a sanitized message, mirroring the WebSocket error contract so
-/// SSE clients get the same codes (`inference_error`, `inference_panic`,
-/// `inference_timeout`, …) instead of one generic string.
-struct StreamError {
-    code: &'static str,
-    message: String,
-}
-
-/// Render one SSE segment-or-error result to the JSON payload string sent in
-/// the `data:` field. Pure (no I/O) so the per-variant error `code`, the
-/// `inference_panic` / `inference_timeout` events, and the partial/final
-/// framing can be unit-tested without a model.
-fn sse_data_payload(
-    result: &Result<gigastt_core::inference::TranscriptSegment, StreamError>,
-) -> String {
-    match result {
-        Ok(seg) => {
-            let ty = if seg.is_final { "final" } else { "partial" };
-            let mut payload = serde_json::json!({
-                "type": ty,
-                "text": seg.text,
-                "timestamp": seg.timestamp,
-                "words": seg.words,
-            });
-            // Same omission contract as the WS segment payload: no words →
-            // no `confidence` key at all.
-            if let Some(confidence) = seg.confidence {
-                payload["confidence"] = confidence.into();
-            }
-            payload.to_string()
-        }
-        Err(err) => serde_json::json!({
-            "type": "error",
-            "message": err.message,
-            "code": err.code,
-        })
-        .to_string(),
-    }
-}
-
-/// Readiness probe response.
-#[derive(Serialize)]
-pub struct ReadinessResponse {
-    /// `"ready"` when the server can accept requests, `"not_ready"` otherwise.
-    pub status: String,
-    /// Number of session triplets currently available for checkout.
-    pub pool_available: usize,
-    /// Total number of session triplets in the pool.
-    pub pool_total: usize,
-    /// Machine-readable reason code when not ready (e.g. `"pool_exhausted"`, `"shutting_down"`).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reason: Option<String>,
-}
-
-/// GET /health — liveness check for monitoring and Docker HEALTHCHECK.
-///
-/// Liveness: stays 200 while the process is alive. It reads only the engine's
-/// static identity (loaded head + effective punctuation/ITN policy) — a cheap,
-/// infallible field read, no pool checkout or I/O — so a client can confirm
-/// *which* model is serving from the same probe it already makes. Pool /
-/// shutdown readiness remains the separate `/ready` probe (see [`readiness`]).
-///
-/// During first-run model download / quantization the listener is served by a
-/// minimal bootstrap responder (see [`super::run_with_config_loading`]) that
-/// reports `model: "loading"`; this handler only runs once the engine is ready.
-pub async fn health(State(state): State<Arc<AppState>>) -> Json<HealthResponse> {
-    let engine = state.engine.load_full();
-    let variant = engine.variant();
-    Json(HealthResponse {
-        status: "ok".into(),
-        model: variant.model_id().into(),
-        variant: variant.as_str().into(),
-        version: env!("CARGO_PKG_VERSION").into(),
-        punctuation: engine.has_punctuator(),
-        itn: engine.has_itn(),
-    })
-}
-
-/// Sample the dedicated batch pool's availability / waiters when one exists
-/// (`--batch-pool-size > 0`). The batch pool has its own FIFO queue, so it can
-/// be saturated while the interactive pool reads healthy; exporting it under
-/// distinct gauges keeps batch-pool exhaustion observable instead of hidden.
-/// No-op when no batch pool was split off.
-pub(crate) fn sample_batch_pool_gauges(reg: &MetricsRegistry, engine: &Engine) {
-    if let Some(ref batch) = engine.batch_pool {
-        reg.gauge_set(
-            "gigastt_batch_pool_available",
-            &[],
-            batch.available() as i64,
-        );
-        reg.gauge_set("gigastt_batch_pool_waiters", &[], batch.waiters() as i64);
-    }
-}
-
-/// GET /ready — readiness probe for k8s and orchestrators.
-pub async fn readiness(State(state): State<Arc<AppState>>) -> Response {
-    let engine = state.engine.load_full();
-    if state.shutdown.is_cancelled() {
-        return (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(ReadinessResponse {
-                status: "not_ready".into(),
-                pool_available: 0,
-                pool_total: engine.pool.total(),
-                reason: Some("shutting_down".into()),
-            }),
-        )
-            .into_response();
-    }
-    let available = engine.pool.available();
-    if let Some(ref reg) = state.metrics_registry {
-        reg.gauge_set("gigastt_pool_available", &[], available as i64);
-        reg.gauge_set("gigastt_pool_waiters", &[], engine.pool.waiters() as i64);
-        sample_batch_pool_gauges(reg, &engine);
-    }
-    if available == 0 {
-        return (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(ReadinessResponse {
-                status: "not_ready".into(),
-                pool_available: 0,
-                pool_total: engine.pool.total(),
-                reason: Some("pool_exhausted".into()),
-            }),
-        )
-            .into_response();
-    }
-    Json(ReadinessResponse {
-        status: "ready".into(),
-        pool_available: available,
-        pool_total: engine.pool.total(),
-        reason: None,
-    })
-    .into_response()
-}
-
-/// GET /v1/models — list loaded models and capabilities.
-pub async fn models(State(state): State<Arc<AppState>>) -> Json<ModelInfo> {
-    let engine = state.engine.load_full();
-    #[cfg(feature = "diarization")]
-    let diarization = engine.has_speaker_encoder();
-    #[cfg(not(feature = "diarization"))]
-    let diarization = false;
-    if let Some(ref reg) = state.metrics_registry {
-        reg.gauge_set(
-            "gigastt_pool_available",
-            &[],
-            engine.pool.available() as i64,
-        );
-        reg.gauge_set("gigastt_pool_waiters", &[], engine.pool.waiters() as i64);
-        sample_batch_pool_gauges(reg, &engine);
-    }
-    let variant = engine.variant();
-    Json(ModelInfo {
-        id: variant.model_id().into(),
-        name: variant.display_name().into(),
-        variant: variant.as_str().into(),
-        version: env!("CARGO_PKG_VERSION").into(),
-        encoder: if engine.is_int8() {
-            "int8".into()
-        } else {
-            "fp32".into()
-        },
-        vocab_size: engine.vocab_size(),
-        sample_rate: 16000,
-        pool_size: engine.pool.total(),
-        pool_available: engine.pool.available(),
-        supported_formats: vec![
-            "wav".into(),
-            "mp3".into(),
-            "m4a".into(),
-            "ogg".into(),
-            "flac".into(),
-        ],
-        supported_rates: super::config::SUPPORTED_RATES.to_vec(),
-        punctuation: engine.has_punctuator(),
-        itn: engine.has_itn(),
-        diarization,
-    })
-}
-
-/// Shared file-transcription pipeline used by `/v1/transcribe` and the
-/// OpenAI-compatible `/v1/audio/transcriptions` alias. Returns the engine
-/// result so each surface can shape its own response envelope.
-async fn run_file_transcription(
-    state: &AppState,
-    body: Bytes,
-    params: &ExportParams,
-) -> Result<gigastt_core::inference::TranscribeResult, ApiError> {
-    if body.is_empty() {
-        return Err(api_error(
-            StatusCode::BAD_REQUEST,
-            "Empty request body",
-            "empty_body",
-        ));
-    }
-
-    // Defence-in-depth: `DefaultBodyLimit` already rejects oversized bodies
-    // before they reach this handler, but a mis-ordered middleware stack or
-    // a `Content-Length`-spoofing client could still deliver too many bytes.
-    // The explicit 413 keeps the REST contract honest and gives clients a
-    // machine-readable `payload_too_large` code alongside the spec-conformant
-    // status. Cheap: `Bytes::len()` is a load, not a walk.
-    let limits = state.limits.load();
-    if body.len() > limits.body_limit_bytes {
-        return Err(api_error(
-            StatusCode::PAYLOAD_TOO_LARGE,
-            "Request body exceeds the configured size limit",
-            "payload_too_large",
-        ));
-    }
-
-    let split_channels = params.channels.as_deref() == Some("split");
-    let request_diarization = params.diarization == Some(true);
-    if split_channels && request_diarization {
-        return Err(api_error(
-            StatusCode::BAD_REQUEST,
-            "channels=split and diarization=true are mutually exclusive",
-            "conflicting_modes",
-        ));
-    }
-
-    // Raw telephony upload (`?codec=`): validate the codec/rate pair up front
-    // so a malformed request 400s before a pool slot is reserved. The actual
-    // decode happens inside the blocking closure below.
-    let raw_codec = resolve_raw_codec(params)?;
-
-    // Snapshot the current engine once at request start; a concurrent hot-reload
-    // swaps the `ArcSwap`, but this request keeps the engine (and its pool) it
-    // began with for its whole lifetime.
-    let engine = state.engine.load_full();
-
-    // Per-request recognition-knob overrides (additive; all absent = the boot
-    // defaults, byte-identical to the pre-feature response). Validate them
-    // *before* checking out a session so an impossible request (a knob turned on
-    // without its resource loaded, or a variant this single-model engine can't
-    // serve) fails fast with a 409 instead of holding a pool triplet.
-    if let Some(requested) = params.variant.as_deref() {
-        // Forward-compat guard only: a single-variant engine can't switch heads,
-        // so a `?variant=X` where X != the loaded variant is a 409. An unknown
-        // token likewise can't match the loaded variant, so it 409s too.
-        let matches = gigastt_core::model::ModelVariant::from_str(requested)
-            .map(|v| v == engine.variant())
-            .unwrap_or(false);
-        if !matches {
-            return Err(api_error(
-                StatusCode::CONFLICT,
-                "Requested model variant is not loaded",
-                "variant_not_loaded",
-            ));
-        }
-    }
-    let overrides = overrides_from_export_params(params);
-    if let Err(e) = engine.validate_overrides(&overrides) {
-        return Err(api_error(StatusCode::CONFLICT, e.message(), e.code()));
-    }
-    let hotwords = hotwords_from_export_params(params);
-    if let Some(ref hw) = hotwords
-        && let Err(e) = engine.validate_hotwords(hw)
-    {
-        return Err(api_error(StatusCode::BAD_REQUEST, e.message(), e.code()));
-    }
-
-    // Checkout a session triplet from the batch pool (blocks if none available)
-    // — this is a long file-transcription job, so it draws from the dedicated
-    // batch pool when one exists (falling back to the interactive pool otherwise)
-    // to avoid starving WebSocket / SSE streaming. Shared with the SSE handler
-    // (`reserve_batch_slot`) so both reserve identically, and before decoding.
-    let mut reservation =
-        reserve_batch_slot(&engine, &limits, state.metrics_registry.as_ref()).await?;
-
-    let inference_start = std::time::Instant::now();
-    let span = tracing::Span::current();
-    let handle = tokio::task::spawn_blocking(move || {
-        let _enter = span.enter();
-        // catch_unwind ensures triplet is returned to pool even on panic
-        let r = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            // Raw telephony upload: decode the headerless byte stream and
-            // re-wrap it as an in-memory WAV so every downstream path (mono,
-            // channels=split, diarized) works unchanged.
-            let body = match raw_codec {
-                Some((codec, rate)) => raw_codec_to_wav(&body, codec, rate)?,
-                None => body,
-            };
-            if split_channels {
-                let channels = gigastt_core::inference::audio::decode_audio_bytes_shared_channels(
-                    body.clone(),
-                )
-                .map_err(|e| gigastt_core::error::GigasttError::InvalidAudio {
-                    reason: format!("{e:#}"),
-                })?;
-                let fallback_reason = match channels.len() {
-                    0 => Some("no channels"),
-                    1 => Some("mono audio"),
-                    2 if gigastt_core::inference::audio::is_dual_mono(&channels) => {
-                        Some("dual-mono audio")
-                    }
-                    n if n > 2 => Some("more than two channels"),
-                    _ => None,
-                };
-                if let Some(reason) = fallback_reason {
-                    tracing::warn!(
-                        "channels=split requested but {reason} detected; falling back to mono transcription"
-                    );
-                    engine.transcribe_bytes_shared(body, &mut reservation)
-                } else {
-                    engine.transcribe_channels(&channels, &mut reservation)
-                }
-            } else if request_diarization {
-                // Diarization is opt-in (`?diarization=true`): only then run the
-                // offline speaker pass. `body` is `axum::body::Bytes` (a
-                // `bytes::Bytes` re-export) so the decode shares the upload buffer.
-                engine.transcribe_bytes_shared_with_overrides_diarized_hotwords(
-                    body,
-                    &mut reservation,
-                    &overrides,
-                    hotwords.as_ref(),
-                )
-            } else {
-                // No diarization requested: byte-identical to the pre-existing
-                // zero-copy path. `overrides` is already validated above.
-                engine.transcribe_bytes_shared_with_overrides_hotwords(
-                    body,
-                    &mut reservation,
-                    &overrides,
-                    hotwords.as_ref(),
-                )
-            }
-        }));
-        match r {
-            Ok(inference_result) => inference_result,
-            Err(_) => {
-                tracing::error!("Panic in REST transcribe — triplet recovered");
-                Err(gigastt_core::error::GigasttError::Inference {
-                    source: anyhow::anyhow!("Inference thread panicked").into(),
-                })
-            }
-        }
-        // reservation dropped here automatically returns the triplet to the pool
-    });
-
-    // Guard the blocking ONNX run with the per-request inference timeout
-    // (`0` disables). `spawn_blocking` can't be cancelled, so the detached task
-    // keeps the triplet and returns the slot to the pool only when the run
-    // finishes; the client gets a typed `inference_timeout` (504) immediately.
-    let inference_timeout_secs = limits.inference_timeout_secs;
-    let result = if inference_timeout_secs == 0 {
-        handle.await
-    } else {
-        match tokio::time::timeout(
-            std::time::Duration::from_secs(inference_timeout_secs),
-            handle,
-        )
-        .await
-        {
-            Ok(r) => r,
-            Err(_elapsed) => {
-                if let Some(ref reg) = state.metrics_registry {
-                    reg.counter_inc("gigastt_inference_timeouts_total", &[], 1);
-                }
-                tracing::error!("REST inference exceeded {inference_timeout_secs}s — aborting");
-                return Err(api_inference_timeout_error());
-            }
-        }
-    };
-    if let Some(ref reg) = state.metrics_registry {
-        reg.histogram_record(
-            "gigastt_inference_duration_seconds",
-            &[],
-            inference_start.elapsed().as_secs_f64(),
-        );
-    }
-
-    match result {
-        Ok(Ok(result)) => Ok(result),
-        Ok(Err(e)) => {
-            tracing::error!("Transcription error: {e}");
-            Err(api_error(
-                StatusCode::UNPROCESSABLE_ENTITY,
-                "Transcription failed. Check audio format.",
-                "transcription_error",
-            ))
-        }
-        Err(e) => {
-            // spawn_blocking task itself failed (e.g., runtime shutdown).
-            // The reservation was dropped inside the closure and the triplet
-            // was returned to the pool automatically.
-            tracing::error!("spawn_blocking join error: {e}");
-            Err(api_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "Internal server error",
-                "internal",
-            ))
-        }
-    }
-}
-
-/// POST /v1/transcribe — upload audio file, get full transcript.
-///
-/// Accepts raw audio body. Supported formats: WAV (including G.711 A-law /
-/// μ-law and G.722 inside WAV), MP3, M4A/AAC, OGG, FLAC — plus headerless
-/// telephony streams when `?codec=pcmu|pcma|g722&sample_rate=N` is given.
-/// Max body size enforced by the axum `DefaultBodyLimit` layer configured
-/// from [`RuntimeLimits::body_limit_bytes`] (default 50 MiB).
-pub async fn transcribe(
-    State(state): State<Arc<AppState>>,
-    Query(params): Query<ExportParams>,
-    body: Bytes,
-) -> Result<Response, ApiError> {
-    let result = run_file_transcription(&state, body, &params).await?;
-    if let Some(rendered) = render_export_response(&result, &params)? {
-        Ok(rendered)
-    } else {
-        // Default JSON response. `?segments=true` adds a cue-grouped
-        // `segments` array (same boundaries as SRT/VTT) alongside the
-        // unchanged top-level `text` / `words` / `duration`; absent
-        // otherwise so existing clients see the exact same shape.
-        let segments = if params.segments.unwrap_or(false) {
-            Some(gigastt_core::export::to_transcript_segments(&result.words))
-        } else {
-            None
-        };
-        Ok(Json(TranscribeResponse {
-            text: result.text,
-            words: result.words,
-            duration: result.duration_s,
-            confidence: result.confidence,
-            segments,
-        })
-        .into_response())
-    }
-}
-
-/// POST /v1/audio/transcriptions — OpenAI-compatible file transcription.
-///
-/// Compatibility surface for clients that speak the OpenAI Audio Transcriptions
-/// API (llama-swap, Hermes Agent, OpenAI SDKs with a custom `base_url`):
-/// `multipart/form-data` with required `file` and optional
-/// `model` / `response_format` / `language` / `timestamp_granularities[]` /
-/// `stream`.
-///
-/// | `response_format` | Body |
-/// |---|---|
-/// | `json` (default) | `{"text":"..."}` |
-/// | `text` | plain text |
-/// | `srt` / `vtt` | captions |
-/// | `verbose_json` | Whisper-style JSON (`task`, `language`, `duration`, `text`, optional `segments`/`words`) |
-///
-/// With `stream=true` (only with `json`/`text`): SSE of
-/// `transcript.text.delta` events, a final `transcript.text.done`, then
-/// `data: [DONE]`. Progressive deltas come from the real chunked encoder path.
-///
-/// Reuses the same inference pipeline as [`transcribe`]. `model` is accepted
-/// and ignored (single loaded head). For diarization, telephony codecs, or
-/// native export knobs use `/v1/transcribe`.
-pub async fn openai_transcriptions(
-    State(state): State<Arc<AppState>>,
-    multipart: Multipart,
-) -> Result<Response, ApiError> {
-    let req = super::openai::parse_openai_multipart(multipart).await?;
-    if req.options.stream {
-        return openai_transcriptions_stream(state, req.file).await;
-    }
-    let result = run_file_transcription(&state, req.file, &ExportParams::default()).await?;
-    Ok(super::openai::render_openai_response(&result, &req.options))
-}
-
-/// OpenAI `stream=true` path: chunked file transcription as SSE transcript events.
-async fn openai_transcriptions_stream(
-    state: Arc<AppState>,
-    body: Bytes,
-) -> Result<Response, ApiError> {
-    // Same early guards as `/v1/transcribe/stream` — empty / oversized body.
-    if body.is_empty() {
-        return Err(api_error(
-            StatusCode::BAD_REQUEST,
-            "Empty request body",
-            "empty_body",
-        ));
-    }
-    let limits = state.limits.load();
-    if body.len() > limits.body_limit_bytes {
-        return Err(api_error(
-            StatusCode::PAYLOAD_TOO_LARGE,
-            "Request body exceeds the configured size limit",
-            "payload_too_large",
-        ));
-    }
-
-    let engine = state.engine.load_full();
-    let mut reservation =
-        reserve_batch_slot(&engine, &limits, state.metrics_registry.as_ref()).await?;
-
-    let samples = tokio::task::spawn_blocking(move || {
-        match std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            gigastt_core::inference::audio::decode_audio_bytes_shared(body)
-        })) {
-            Ok(inner) => inner,
-            Err(_) => {
-                tracing::error!("Panic in OpenAI SSE audio decode — treated as decode error");
-                Err(anyhow::anyhow!("Audio decode thread panicked"))
-            }
-        }
-    })
-    .await
-    .map_err(|e| {
-        tracing::error!("spawn_blocking join error: {e}");
-        api_error(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "Internal server error",
-            "internal",
-        )
-    })?
-    .map_err(|e| {
-        tracing::error!("Audio decode error: {e:#}");
-        api_error(
-            StatusCode::UNPROCESSABLE_ENTITY,
-            "Failed to decode audio file. Check format (WAV, MP3, M4A, OGG, FLAC supported).",
-            "invalid_audio",
-        )
-    })?;
-
-    // Channel of pre-rendered SSE `data:` payloads (JSON events or `[DONE]`).
-    let (tx, rx) = tokio::sync::mpsc::channel::<String>(32);
-
-    let cancel = state.shutdown.clone();
-    let tracker = state.tracker.clone();
-    let span = tracing::Span::current();
-    tracker.spawn_blocking(move || {
-        let _enter = span.enter();
-        use super::openai::{OpenAIStreamAssembler, sse_delta_payload, sse_done_payload};
-
-        let send = |payload: String| -> bool { tx.blocking_send(payload).is_ok() };
-
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let mut stream_state = engine.create_state(false);
-            let mut asm = OpenAIStreamAssembler::new();
-            let chunk_size = 16000; // 1 s @ 16 kHz
-
-            for chunk in samples.chunks(chunk_size) {
-                if cancel.is_cancelled() {
-                    tracing::info!("OpenAI SSE transcription cancelled by shutdown");
-                    return;
-                }
-                match engine.process_chunk(chunk, &mut stream_state, &mut reservation) {
-                    Ok(segs) => {
-                        for seg in segs {
-                            if let Some(delta) = asm.push_segment(&seg.text, seg.is_final)
-                                && !send(sse_delta_payload(&delta))
-                            {
-                                return; // client gone
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        tracing::error!("OpenAI SSE transcription error: {e}");
-                        // Surface a final done with whatever we have so clients
-                        // do not hang; OpenAI stream errors are not standardized.
-                        let _ = send(sse_done_payload(asm.text()));
-                        let _ = send("[DONE]".into());
-                        return;
-                    }
-                }
-            }
-
-            if let Some(seg) = engine.finish_stream(&mut stream_state, &mut reservation)
-                && let Some(delta) = asm.push_segment(&seg.text, seg.is_final)
-            {
-                let _ = send(sse_delta_payload(&delta));
-            }
-
-            let _ = send(sse_done_payload(asm.text()));
-            let _ = send("[DONE]".into());
-        }));
-
-        if result.is_err() {
-            tracing::error!("Panic in OpenAI SSE inference task — triplet recovered");
-            let _ = send(sse_done_payload(""));
-            let _ = send("[DONE]".into());
-        }
-        // reservation dropped → pool
-    });
-
-    let stream = tokio_stream::wrappers::ReceiverStream::new(rx)
-        .map(|data| Ok::<_, std::convert::Infallible>(super::openai::sse_event_data(data)));
-
-    Ok(Sse::new(stream)
-        .keep_alive(
-            KeepAlive::new()
-                .interval(std::time::Duration::from_secs(15))
-                .text(""),
-        )
-        .into_response())
-}
-
-/// POST /v1/transcribe/stream — upload audio file, get SSE stream of partial/final results.
-///
-/// Real streaming: audio is processed chunk-by-chunk inside `spawn_blocking`,
-/// and segments are sent to the SSE stream via an mpsc channel as they are produced.
-pub async fn transcribe_stream(
-    State(state): State<Arc<AppState>>,
-    body: Bytes,
-) -> Result<Sse<impl Stream<Item = Result<Event, std::convert::Infallible>>>, ApiError> {
-    if body.is_empty() {
-        return Err(api_error(
-            StatusCode::BAD_REQUEST,
-            "Empty request body",
-            "empty_body",
-        ));
-    }
-
-    // Defence-in-depth early reject; matches `/v1/transcribe` — see that
-    // handler for the rationale.
-    let limits = state.limits.load();
-    if body.len() > limits.body_limit_bytes {
-        return Err(api_error(
-            StatusCode::PAYLOAD_TOO_LARGE,
-            "Request body exceeds the configured size limit",
-            "payload_too_large",
-        ));
-    }
-
-    // Checkout a session triplet from the batch pool BEFORE decoding. SSE file
-    // transcription decodes and transcribes the *entire* upload (holding the
-    // triplet for the whole file), so it is a batch workload, not interactive
-    // streaming. Reserving first — via the same `reserve_batch_slot` the
-    // synchronous `/v1/transcribe` uses — is load-bearing: it caps the number of
-    // *concurrent decodes* at the pool size, so a burst of large (compressed)
-    // uploads can't each expand into a full f32 PCM buffer at once and exhaust
-    // memory. A saturated pool yields 503 / `Retry-After` here, before any decode.
-    // Snapshot the current engine once; a concurrent hot-reload swap only affects
-    // later requests, so this stream rides the pool it started on.
-    let engine = state.engine.load_full();
-    let mut reservation =
-        reserve_batch_slot(&engine, &limits, state.metrics_registry.as_ref()).await?;
-
-    // Decode audio now that a slot is reserved (in spawn_blocking since symphonia
-    // is blocking). Holding the reservation across the decode is what bounds the
-    // number of concurrent PCM expansions to the pool size. `body` is
-    // `axum::body::Bytes`, so the move into the blocking closure is a refcount
-    // bump and `decode_audio_bytes_shared` reads the upload buffer in place. On a
-    // decode error the early `?` return drops `reservation`, returning the
-    // triplet to the pool.
-    let samples = tokio::task::spawn_blocking(move || {
-        // catch_unwind mirrors the REST handler: a panic inside the blocking
-        // decode (e.g. a crafted container that trips an upstream arithmetic
-        // panic) is absorbed and surfaced as a normal decode error instead of a
-        // `JoinError`, so the SSE path returns a clean 422 rather than a 500.
-        match std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            gigastt_core::inference::audio::decode_audio_bytes_shared(body)
-        })) {
-            Ok(inner) => inner,
-            Err(_) => {
-                tracing::error!("Panic in SSE audio decode — treated as decode error");
-                Err(anyhow::anyhow!("Audio decode thread panicked"))
-            }
-        }
-    })
-    .await
-    .map_err(|e| {
-        tracing::error!("spawn_blocking join error: {e}");
-        api_error(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "Internal server error",
-            "internal",
-        )
-    })?
-    .map_err(|e| {
-        tracing::error!("Audio decode error: {e:#}");
-        api_error(
-            StatusCode::UNPROCESSABLE_ENTITY,
-            "Failed to decode audio file. Check format (WAV, MP3, M4A, OGG, FLAC supported).",
-            "invalid_audio",
-        )
-    })?;
-
-    // Create mpsc channel for streaming segments from the inference task to SSE.
-    let (tx, rx) = tokio::sync::mpsc::channel::<
-        Result<gigastt_core::inference::TranscriptSegment, StreamError>,
-    >(16);
-
-    // The axum handler future has already returned by the time the SSE stream
-    // starts flowing, so `with_graceful_shutdown` can't observe this task. Clone
-    // the shutdown token and check it before every chunk so SIGTERM during a
-    // long transcription drops cleanly.
-    //
-    // The whole file is transcribed in one blocking task, streaming each 1 s
-    // chunk's segments out as they are produced. Each `process_chunk` is a small
-    // bounded unit of work, so unlike the single-shot REST path it is not
-    // wrapped by the per-request inference timeout; liveness on shutdown is
-    // handled by the per-chunk cancellation check.
-    let cancel = state.shutdown.clone();
-    let tracker = state.tracker.clone();
-    let span = tracing::Span::current();
-    tracker.spawn_blocking(move || {
-        let _enter = span.enter();
-        // catch_unwind ensures the triplet is returned to the pool even on panic.
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let mut stream_state = engine.create_state(false);
-            let chunk_size = 16000; // 1 second at 16 kHz
-
-            for chunk in samples.chunks(chunk_size) {
-                if cancel.is_cancelled() {
-                    tracing::info!("SSE transcription cancelled by shutdown");
-                    return;
-                }
-                match engine.process_chunk(chunk, &mut stream_state, &mut reservation) {
-                    Ok(segs) => {
-                        for seg in segs {
-                            if tx.blocking_send(Ok(seg)).is_err() {
-                                // Receiver dropped (client disconnected).
-                                return;
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        let _ = tx.blocking_send(Err(StreamError {
-                            code: e.code(),
-                            message: "Transcription failed. Please check audio format.".into(),
-                        }));
-                        return;
-                    }
-                }
-            }
-
-            // Final decode of the sub-stride remainder, then flush — best-effort;
-            // always emit so SSE clients receive a clean end-of-stream marker.
-            if let Some(seg) = engine.finish_stream(&mut stream_state, &mut reservation) {
-                let _ = tx.blocking_send(Ok(seg));
-            }
-        }));
-
-        if result.is_err() {
-            tracing::error!("Panic in SSE inference task — triplet recovered");
-            // Mirror the WebSocket contract: surface a distinct `inference_panic`
-            // code instead of ending the stream silently.
-            let _ = tx.blocking_send(Err(StreamError {
-                code: "inference_panic",
-                message: "Inference failed unexpectedly.".into(),
-            }));
-        }
-        // reservation dropped here automatically returns the triplet to the pool
-    });
-
-    // Convert receiver to SSE stream.
-    let stream = tokio_stream::wrappers::ReceiverStream::new(rx)
-        .map(|result| Ok(Event::default().data(sse_data_payload(&result))));
-
-    // Explicit keep-alive: send a comment (`: \n\n`) every 15 s so nginx / ALB
-    // do not close the connection during long transcriptions.
-    Ok(Sse::new(stream).keep_alive(
-        KeepAlive::new()
-            .interval(std::time::Duration::from_secs(15))
-            .text(""),
-    ))
-}
-
-/// Whether the reload endpoint should accept a request from `peer`.
-///
-/// Model reload is an administrative, machine-local action: it must stay
-/// reachable only from the loopback interface even under `--bind-all` /
-/// `--cors-allow-any`, which would otherwise widen `origin_middleware` (the only
-/// other gate). Pure so the loopback decision can be unit-tested without a model
-/// or a live socket.
-fn peer_is_loopback(peer: &std::net::SocketAddr) -> bool {
-    peer.ip().is_loopback()
-}
-
-/// POST /v1/admin/reload — rebuild the inference engine from the boot recipe and
-/// atomically swap it in without a restart.
-///
-/// Strictly loopback-only (checked here, not just via the origin middleware),
-/// serialized by a mutex so two reloads can't race, and fail-safe: a build error
-/// leaves the currently-serving engine untouched. The new engine is warmed
-/// before the swap so the first post-swap request doesn't pay the cold cost.
-/// In-flight requests keep the engine they started on and finish against its
-/// pool; the old engine drops when its last reference goes.
-pub async fn reload(
-    axum::extract::ConnectInfo(peer): axum::extract::ConnectInfo<std::net::SocketAddr>,
-    State(state): State<Arc<AppState>>,
-) -> Response {
-    // Gotcha #2: enforce loopback here so reload stays local even when
-    // `origin_middleware` has been widened by `--bind-all` / `--cors-allow-any`
-    // or a caller omits the Origin header.
-    if !peer_is_loopback(&peer) {
-        tracing::warn!(peer = %peer, "Rejecting non-loopback model reload request");
-        return (
-            StatusCode::FORBIDDEN,
-            Json(serde_json::json!({
-                "error": "Model reload is only available from loopback",
-                "code": "loopback_only",
-            })),
-        )
-            .into_response();
-    }
-
-    let Some(builder) = state.engine_builder.clone() else {
-        return (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(serde_json::json!({
-                "error": "Model reload is not available on this server",
-                "code": "reload_unsupported",
-            })),
-        )
-            .into_response();
-    };
-
-    // Serialize reloads: the loser of the race gets 409 rather than queueing, so
-    // an operator hammering the endpoint can't stack up concurrent rebuilds.
-    let _reload_guard = match state.reload_lock.try_lock() {
-        Ok(guard) => guard,
-        Err(_) => {
-            return (
-                StatusCode::CONFLICT,
-                Json(serde_json::json!({
-                    "error": "A model reload is already in progress",
-                    "code": "reload_in_progress",
-                })),
-            )
-                .into_response();
-        }
-    };
-
-    tracing::info!(peer = %peer, "Model reload requested — rebuilding engine");
-
-    // Build the new engine off the request path (ONNX session load is blocking).
-    let build = tokio::task::spawn_blocking(move || builder()).await;
-
-    let new_engine = match build {
-        Ok(Ok(engine)) => engine,
-        Ok(Err(e)) => {
-            // Keep the old engine untouched. Log the detail, return a sanitized
-            // message (no path / model leakage) matching the internal-error policy.
-            tracing::error!("Model reload build failed: {e:#}");
-            return (
-                StatusCode::SERVICE_UNAVAILABLE,
-                Json(serde_json::json!({
-                    "error": "Model reload failed; the previous model is still serving",
-                    "code": "reload_failed",
-                })),
-            )
-                .into_response();
-        }
-        Err(join_err) => {
-            tracing::error!("Model reload build task panicked: {join_err}");
-            return (
-                StatusCode::SERVICE_UNAVAILABLE,
-                Json(serde_json::json!({
-                    "error": "Model reload failed; the previous model is still serving",
-                    "code": "reload_failed",
-                })),
-            )
-                .into_response();
-        }
-    };
-
-    // Warm the fresh engine BEFORE swapping so the first post-swap request
-    // doesn't pay the EP-compile / first-allocation cost. A warmup failure is
-    // non-fatal (mirrors boot): the engine already fell back to CPU internally.
-    let new_engine = match tokio::task::spawn_blocking(move || {
-        if let Err(e) = new_engine.warmup() {
-            tracing::warn!("Reloaded engine warmup failed (swapping anyway): {e:#}");
-        }
-        new_engine
-    })
-    .await
-    {
-        Ok(engine) => engine,
-        Err(join_err) => {
-            tracing::error!("Model reload warmup task panicked: {join_err}");
-            return (
-                StatusCode::SERVICE_UNAVAILABLE,
-                Json(serde_json::json!({
-                    "error": "Model reload failed; the previous model is still serving",
-                    "code": "reload_failed",
-                })),
-            )
-                .into_response();
-        }
-    };
-
-    let variant = new_engine.variant();
-    let encoder = if new_engine.is_int8() { "int8" } else { "fp32" };
-
-    // Atomic swap. In-flight requests holding the old `Arc<Engine>` finish
-    // against the old pool; it drops when its last reference goes. We do NOT
-    // close the old pool — that would strand in-flight work.
-    state.engine.store(Arc::new(new_engine));
-    tracing::info!(
-        variant = variant.as_str(),
-        encoder,
-        "Model reloaded and swapped"
-    );
-
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({
-            "reloaded": true,
-            "variant": variant.as_str(),
-            "encoder": encoder,
-        })),
-    )
-        .into_response()
-}
-
-/// POST /v1/jobs response.
-#[derive(Serialize)]
-pub struct JobSubmitResponse {
-    pub job_id: String,
-    pub status: &'static str,
-    pub created_at: f64,
-}
-
-/// Return the job server state or a 404 if the async job API is disabled.
-#[allow(clippy::result_large_err)]
-fn require_jobs(state: &AppState) -> Result<&JobServerState, ApiError> {
-    state.jobs.as_ref().ok_or_else(|| {
-        api_error(
-            StatusCode::NOT_FOUND,
-            "Job API is not enabled",
-            "jobs_disabled",
-        )
-    })
-}
-
-/// Fetch a job by id, mapping store errors to the standard HTTP responses.
-async fn load_job(store: &dyn JobStore, id: &str) -> Result<super::jobs::Job, ApiError> {
-    match store.get(id).await {
-        Ok(Some(job)) => Ok(job),
-        Ok(None) => Err(api_error(
-            StatusCode::NOT_FOUND,
-            "Job not found",
-            "job_not_found",
-        )),
-        Err(e) => {
-            tracing::error!("Failed to get job {id}: {e:#}");
-            Err(api_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "Failed to read job status",
-                "internal",
-            ))
-        }
-    }
-}
-
-/// POST /v1/jobs — enqueue a long audio file for asynchronous transcription.
-///
-/// Accepts the same body and query parameters as `/v1/transcribe`. Returns 202
-/// with the job id; use `GET /v1/jobs/{id}` to poll and
-/// `GET /v1/jobs/{id}/result` to fetch the finished transcript.
-pub async fn submit_job(
-    State(state): State<Arc<AppState>>,
-    Query(params): Query<ExportParams>,
-    body: Bytes,
-) -> Result<Response, ApiError> {
-    let jobs = require_jobs(&state)?;
-    let limits = state.limits.load();
-    if body.is_empty() {
-        return Err(api_error(
-            StatusCode::BAD_REQUEST,
-            "Empty request body",
-            "empty_body",
-        ));
-    }
-    if body.len() > limits.body_limit_bytes {
-        return Err(api_error(
-            StatusCode::PAYLOAD_TOO_LARGE,
-            "Request body exceeds the configured size limit",
-            "payload_too_large",
-        ));
-    }
-    if jobs.store.is_full().await {
-        return Err((
-            StatusCode::TOO_MANY_REQUESTS,
-            [(
-                header::RETRY_AFTER,
-                pool_retry_after_secs(&limits).to_string(),
-            )],
-            Json(serde_json::json!({
-                "error": "Job queue is full",
-                "code": "queue_full",
-                "retry_after_ms": pool_retry_after_ms(&limits),
-            })),
-        )
-            .into_response());
-    }
-    let split_channels = params.channels.as_deref() == Some("split");
-    let request_diarization = params.diarization == Some(true);
-    if split_channels && request_diarization {
-        return Err(api_error(
-            StatusCode::BAD_REQUEST,
-            "channels=split and diarization=true are mutually exclusive",
-            "conflicting_modes",
-        ));
-    }
-
-    let job = super::jobs::Job::queued(body, params);
-    let created_at = job.created_at;
-    let id = match jobs.store.create(job).await {
-        Ok(id) => id,
-        Err(e) => {
-            tracing::error!("Failed to create job: {e:#}");
-            return Err(api_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "Failed to enqueue job",
-                "internal",
-            ));
-        }
-    };
-    Ok((
-        StatusCode::ACCEPTED,
-        Json(JobSubmitResponse {
-            job_id: id,
-            status: "queued",
-            created_at,
-        }),
-    )
-        .into_response())
-}
-
-/// GET /v1/jobs/{id} — poll job status and progress.
-pub async fn get_job(
-    State(state): State<Arc<AppState>>,
-    axum::extract::Path(id): axum::extract::Path<String>,
-) -> Result<Response, ApiError> {
-    let jobs = require_jobs(&state)?;
-    let job = load_job(&*jobs.store, &id).await?;
-    Ok(Json(super::jobs::job_status_response(&job)).into_response())
-}
-
-/// GET /v1/jobs/{id}/result — fetch the finished transcription.
-pub async fn get_job_result(
-    State(state): State<Arc<AppState>>,
-    axum::extract::Path(id): axum::extract::Path<String>,
-) -> Result<Response, ApiError> {
-    let jobs = require_jobs(&state)?;
-    let job = load_job(&*jobs.store, &id).await?;
-    if !matches!(job.status, JobStatus::Done) {
-        return Err(api_error(
-            StatusCode::CONFLICT,
-            "Job is not finished",
-            "job_not_finished",
-        ));
-    }
-    let Some(result) = job.result else {
-        tracing::error!(job_id = %id, "Done job is missing result");
-        return Err(api_error(
-            StatusCode::INTERNAL_SERVER_ERROR,
-            "Job result is missing",
-            "internal",
-        ));
-    };
-    if let Some(rendered) = render_export_response(&result, &job.params)? {
-        Ok(rendered)
-    } else {
-        let segments = if job.params.segments.unwrap_or(false) {
-            Some(gigastt_core::export::to_transcript_segments(&result.words))
-        } else {
-            None
-        };
-        Ok(Json(TranscribeResponse {
-            text: result.text,
-            words: result.words,
-            duration: result.duration_s,
-            confidence: result.confidence,
-            segments,
-        })
-        .into_response())
-    }
-}
-
-/// DELETE /v1/jobs/{id} — cancel a queued or processing job.
-pub async fn cancel_job(
-    State(state): State<Arc<AppState>>,
-    axum::extract::Path(id): axum::extract::Path<String>,
-) -> Result<Response, ApiError> {
-    let jobs = require_jobs(&state)?;
-    let job = load_job(&*jobs.store, &id).await?;
-    if !matches!(job.status, JobStatus::Queued | JobStatus::Processing) {
-        return Err(api_error(
-            StatusCode::CONFLICT,
-            "Job cannot be cancelled",
-            "job_not_cancellable",
-        ));
-    }
-    let _ = jobs
-        .store
-        .update(
-            &id,
-            Box::new(|j| {
-                if matches!(j.status, JobStatus::Queued | JobStatus::Processing) {
-                    j.status = JobStatus::Cancelled;
-                }
-            }),
-        )
-        .await;
-    super::jobs::broadcast_event(&*jobs.store, &id, JobEvent::Cancelled).await;
-    Ok(StatusCode::NO_CONTENT.into_response())
-}
-
-/// GET /v1/jobs/{id}/events — SSE stream of progress / done / failed / cancelled.
-pub async fn job_events(
-    State(state): State<Arc<AppState>>,
-    axum::extract::Path(id): axum::extract::Path<String>,
-) -> Result<Sse<impl Stream<Item = Result<Event, std::convert::Infallible>>>, ApiError> {
-    let jobs = require_jobs(&state)?;
-    let job = load_job(&*jobs.store, &id).await?;
-
-    let (tx, rx) = tokio::sync::mpsc::unbounded_channel::<JobEvent>();
-    if job.status.is_terminal() {
-        let event = match job.status {
-            JobStatus::Done => JobEvent::Done,
-            JobStatus::Failed => JobEvent::Failed {
-                error: job
-                    .error
-                    .clone()
-                    .unwrap_or_else(|| "Transcription failed.".into()),
-            },
-            JobStatus::Cancelled => JobEvent::Cancelled,
-            _ => unreachable!(),
-        };
-        let _ = tx.send(event);
-    } else {
-        let _ = jobs
-            .store
-            .update(
-                &id,
-                Box::new(move |j| {
-                    j.subscribe(tx);
-                }),
-            )
-            .await;
-    }
-
-    let stream = tokio_stream::wrappers::UnboundedReceiverStream::new(rx)
-        .map(|event| Ok(Event::default().data(serde_json::to_string(&event).unwrap_or_default())));
-
-    Ok(Sse::new(stream).keep_alive(
-        KeepAlive::new()
-            .interval(std::time::Duration::from_secs(15))
-            .text(""),
-    ))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_health_response_serialization() {
-        let resp = HealthResponse {
-            status: "ok".into(),
-            model: "gigaam-v3-rnnt".into(),
-            variant: "rnnt".into(),
-            version: "0.3.0".into(),
-            punctuation: true,
-            itn: true,
-        };
-        let json = serde_json::to_string(&resp).unwrap();
-        let v: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert_eq!(v["status"], "ok");
-        assert_eq!(v["model"], "gigaam-v3-rnnt");
-        assert_eq!(v["variant"], "rnnt");
-        assert_eq!(v["punctuation"], true);
-        assert_eq!(v["itn"], true);
-    }
-
-    #[test]
-    fn test_transcribe_response_serialization() {
-        let resp = TranscribeResponse {
-            text: "hello".into(),
-            words: vec![],
-
-            duration: 1.5,
-            confidence: None,
-            segments: None,
-        };
-        let json = serde_json::to_string(&resp).unwrap();
-        let v: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert_eq!(v["text"], "hello");
-        assert_eq!(v["duration"], 1.5);
-    }
-
-    #[test]
-    fn test_readiness_response_ready_serialization() {
-        let resp = ReadinessResponse {
-            status: "ready".into(),
-            pool_available: 3,
-            pool_total: 4,
-            reason: None,
-        };
-        let json = serde_json::to_value(&resp).unwrap();
-        assert_eq!(json["status"], "ready");
-        assert_eq!(json["pool_available"], 3);
-        assert_eq!(json["pool_total"], 4);
-        assert!(json.get("reason").is_none() || json["reason"].is_null());
-    }
-
-    #[test]
-    fn test_readiness_response_not_ready_serialization() {
-        let resp = ReadinessResponse {
-            status: "not_ready".into(),
-            pool_available: 0,
-            pool_total: 4,
-            reason: Some("pool_exhausted".into()),
-        };
-        let json = serde_json::to_value(&resp).unwrap();
-        assert_eq!(json["status"], "not_ready");
-        assert_eq!(json["reason"], "pool_exhausted");
-    }
-
-    #[tokio::test]
-    async fn test_api_error_basic() {
-        let resp = api_error(StatusCode::BAD_REQUEST, "bad request", "bad_request");
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::BAD_REQUEST);
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["error"], "bad request");
-        assert_eq!(v["code"], "bad_request");
-    }
-
-    #[tokio::test]
-    async fn test_override_conflict_error_mapping() {
-        // The per-request-knob 409s reuse the shared `api_error` machinery, so
-        // they must carry StatusCode::CONFLICT and the stable code an operator's
-        // client keys off. Drive it via the same `OverrideError::{code,message}`
-        // the handler maps, plus the standalone `variant_not_loaded` guard.
-        use gigastt_core::inference::OverrideError;
-        for err in [
-            OverrideError::VadNotLoaded,
-            OverrideError::PunctuationNotAvailable,
-        ] {
-            let resp = api_error(StatusCode::CONFLICT, err.message(), err.code());
-            let (parts, body) = resp.into_parts();
-            assert_eq!(parts.status, StatusCode::CONFLICT);
-            let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-            let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-            assert_eq!(v["code"], err.code());
-            assert!(v["error"].as_str().is_some_and(|s| !s.is_empty()));
-        }
-        assert_eq!(OverrideError::VadNotLoaded.code(), "vad_not_loaded");
-        assert_eq!(
-            OverrideError::PunctuationNotAvailable.code(),
-            "punctuation_not_available"
-        );
-
-        // Hotword DoS limit violations map to 400 (not 409).
-        use gigastt_core::inference::HotwordError;
-        for err in [HotwordError::TooManyHotwords, HotwordError::PhraseTooLong] {
-            let resp = api_error(StatusCode::BAD_REQUEST, err.message(), err.code());
-            let (parts, body) = resp.into_parts();
-            assert_eq!(parts.status, StatusCode::BAD_REQUEST);
-            let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-            let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-            assert_eq!(v["code"], err.code());
-        }
-
-        // The variant guard is a standalone literal (no engine needed to check
-        // the code/status contract it emits).
-        let resp = api_error(
-            StatusCode::CONFLICT,
-            "Requested model variant is not loaded",
-            "variant_not_loaded",
-        );
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::CONFLICT);
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "variant_not_loaded");
-    }
-
-    #[tokio::test]
-    async fn test_api_timeout_error_includes_retry_after() {
-        let limits = RuntimeLimits::default();
-        let resp = api_timeout_error(&limits);
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::SERVICE_UNAVAILABLE);
-        assert_eq!(
-            parts
-                .headers
-                .get(header::RETRY_AFTER)
-                .unwrap()
-                .to_str()
-                .unwrap(),
-            pool_retry_after_secs(&limits).to_string()
-        );
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "timeout");
-        assert_eq!(v["retry_after_ms"], pool_retry_after_ms(&limits));
-    }
-
-    #[tokio::test]
-    async fn test_api_pool_closed_error_no_retry() {
-        let resp = api_pool_closed_error();
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::SERVICE_UNAVAILABLE);
-        assert!(parts.headers.get(header::RETRY_AFTER).is_none());
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "pool_closed");
-        assert!(v.get("retry_after_ms").is_none());
-    }
-
-    #[tokio::test]
-    async fn test_api_inference_timeout_error_is_504() {
-        let resp = api_inference_timeout_error();
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::GATEWAY_TIMEOUT);
-        // A wedged run would just time out again, so no Retry-After hint.
-        assert!(parts.headers.get(header::RETRY_AFTER).is_none());
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "inference_timeout");
-    }
-
-    #[test]
-    fn test_resolve_raw_codec_absent_is_none() {
-        let params = ExportParams::default();
-        assert!(resolve_raw_codec(&params).unwrap().is_none());
-    }
-
-    #[test]
-    fn test_resolve_raw_codec_valid_pairs() {
-        for (name, rate) in [
-            ("pcmu", 8000),
-            ("ulaw", 8000),
-            ("pcma", 8000),
-            ("alaw", 16000),
-            ("g722", 8000),
-            ("g722", 16000),
-        ] {
-            let params = ExportParams {
-                codec: Some(name.into()),
-                sample_rate: Some(rate),
-                ..Default::default()
-            };
-            assert!(
-                resolve_raw_codec(&params).unwrap().is_some(),
-                "{name}@{rate} must resolve"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_resolve_raw_codec_unknown_codec_is_400() {
-        let params = ExportParams {
-            codec: Some("g729".into()),
-            sample_rate: Some(8000),
-            ..Default::default()
-        };
-        let resp = resolve_raw_codec(&params).unwrap_err();
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::BAD_REQUEST);
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "unsupported_codec");
-    }
-
-    #[tokio::test]
-    async fn test_resolve_raw_codec_missing_sample_rate_is_400() {
-        let params = ExportParams {
-            codec: Some("pcmu".into()),
-            sample_rate: None,
-            ..Default::default()
-        };
-        let resp = resolve_raw_codec(&params).unwrap_err();
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::BAD_REQUEST);
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "invalid_sample_rate");
-    }
-
-    #[tokio::test]
-    async fn test_resolve_raw_codec_bad_sample_rate_is_400() {
-        for (name, rate) in [("pcmu", 4000), ("g722", 44100)] {
-            let params = ExportParams {
-                codec: Some(name.into()),
-                sample_rate: Some(rate),
-                ..Default::default()
-            };
-            let resp = resolve_raw_codec(&params).unwrap_err();
-            let (parts, body) = resp.into_parts();
-            assert_eq!(parts.status, StatusCode::BAD_REQUEST, "{name}@{rate}");
-            let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-            let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-            assert_eq!(v["code"], "invalid_sample_rate", "{name}@{rate}");
-        }
-    }
-
-    #[test]
-    fn test_raw_codec_to_wav_produces_decodable_wav() {
-        // μ-law silence (0xFF ≈ 0) re-wraps into a WAV the standard pipeline
-        // accepts: the full raw→16kHz-WAV transform without a model.
-        let raw = vec![0xFFu8; 8000]; // 1 s of μ-law silence at 8 kHz
-        let wav = raw_codec_to_wav(
-            &raw,
-            gigastt_core::inference::audio::TelephonyCodec::Pcmu,
-            8000,
-        )
-        .unwrap();
-        let samples = gigastt_core::inference::audio::decode_audio_bytes_shared(wav).unwrap();
-        assert!(
-            samples.len() > 12_000 && samples.len() <= 16_000,
-            "expected ~1 s at 16 kHz, got {}",
-            samples.len()
-        );
-        assert!(
-            samples.iter().all(|s| s.abs() < 0.01),
-            "μ-law silence must decode to near-silence"
-        );
-    }
-
-    #[test]
-    fn test_raw_codec_to_wav_rejects_bad_input() {
-        let result = raw_codec_to_wav(
-            &[],
-            gigastt_core::inference::audio::TelephonyCodec::Pcmu,
-            8000,
-        );
-        assert!(result.is_err(), "empty raw payload must error");
-    }
-
-    #[test]
-    fn test_export_params_parse_codec_query() {
-        // The query string is how REST clients pass the pair; pin it through
-        // axum's Query extractor itself (the exact extraction path the
-        // handler uses).
-        let uri: axum::http::Uri = "http://localhost/v1/transcribe?codec=pcmu&sample_rate=8000"
-            .parse()
-            .unwrap();
-        let axum::extract::Query(params) = axum::extract::Query::<ExportParams>::try_from_uri(&uri)
-            .expect("codec query must parse");
-        assert_eq!(params.codec.as_deref(), Some("pcmu"));
-        assert_eq!(params.sample_rate, Some(8000));
-    }
-
-    #[test]
-    fn test_sse_data_payload_preserves_error_codes() {
-        // Per-variant code is preserved (not collapsed to a generic string),
-        // including the distinct inference_panic / inference_timeout events.
-        for code in [
-            "invalid_audio",
-            "inference_error",
-            "inference_panic",
-            "inference_timeout",
-        ] {
-            let payload = sse_data_payload(&Err(StreamError {
-                code,
-                message: "sanitized".into(),
-            }));
-            let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-            assert_eq!(v["type"], "error");
-            assert_eq!(v["code"], code);
-            assert_eq!(v["message"], "sanitized");
-        }
-    }
-
-    #[test]
-    fn test_sse_data_payload_segment_framing() {
-        // A final segment renders as type "final"; a non-final one as "partial".
-        let seg = gigastt_core::inference::TranscriptSegment::empty_final();
-        let final_payload = sse_data_payload(&Ok(seg));
-        let v: serde_json::Value = serde_json::from_str(&final_payload).unwrap();
-        assert_eq!(v["type"], "final");
-
-        let mut partial = gigastt_core::inference::TranscriptSegment::empty_final();
-        partial.is_final = false;
-        let partial_payload = sse_data_payload(&Ok(partial));
-        let v: serde_json::Value = serde_json::from_str(&partial_payload).unwrap();
-        assert_eq!(v["type"], "partial");
-    }
-
-    #[test]
-    fn test_sse_data_payload_confidence_present_only_when_some() {
-        // A segment with words carries the aggregate; an empty one omits the
-        // key entirely, matching the WS payload contract.
-        let mut seg = gigastt_core::inference::TranscriptSegment::empty_final();
-        seg.confidence = Some(0.85);
-        let payload = sse_data_payload(&Ok(seg));
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        let c = v["confidence"].as_f64().expect("numeric confidence");
-        assert!((c - 0.85).abs() < 1e-6, "got {c}");
-
-        let empty = gigastt_core::inference::TranscriptSegment::empty_final();
-        let payload = sse_data_payload(&Ok(empty));
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        assert!(v.get("confidence").is_none());
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_readiness_when_shutdown_cancelled() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        state.shutdown.cancel();
-        let resp = readiness(State(state)).await;
-        assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE);
-        let bytes = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["status"], "not_ready");
-        assert_eq!(v["reason"], "shutting_down");
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_readiness_when_pool_exhausted() {
-        let engine = fresh_engine();
-        let _guards: Vec<_> = (0..engine.pool.total())
-            .map(|_| engine.pool.checkout_blocking().unwrap())
-            .collect();
-        let state = Arc::new(AppState {
-            engine: engine_swap(engine),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let resp = readiness(State(state)).await;
-        assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE);
-        let bytes = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["status"], "not_ready");
-        assert_eq!(v["reason"], "pool_exhausted");
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_payload_too_large() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits {
-                body_limit_bytes: 10,
-                ..RuntimeLimits::default()
-            })),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let body = Bytes::from(vec![0u8; 100]);
-        let result = transcribe(State(state), Query(ExportParams::default()), body).await;
-        match result {
-            Err(resp) => assert_eq!(resp.status(), StatusCode::PAYLOAD_TOO_LARGE),
-            Ok(_) => panic!("expected payload_too_large error"),
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_channels_split_diarization_conflict_returns_400() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let params = ExportParams {
-            channels: Some("split".into()),
-            diarization: Some(true),
-            ..ExportParams::default()
-        };
-        let resp = transcribe(State(state), Query(params), minimal_wav())
-            .await
-            .unwrap_err();
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-        let bytes = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "conflicting_modes");
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_models_with_metrics() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: Some(Arc::new(MetricsRegistry::new())),
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let resp = models(State(state)).await;
-        let json = serde_json::to_value(&*resp).unwrap();
-        // The id reflects the head actually loaded on disk (rnnt or e2e_rnnt),
-        // not a hardcoded literal, so assert the stable shape instead.
-        let id = json["id"].as_str().unwrap();
-        assert!(
-            id == "gigaam-v3-rnnt" || id == "gigaam-v3-e2e-rnnt",
-            "unexpected model id: {id}"
-        );
-        assert_eq!(
-            json["variant"],
-            if id.contains("e2e") {
-                "e2e_rnnt"
-            } else {
-                "rnnt"
-            }
-        );
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_readiness_with_metrics() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(fresh_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: Some(Arc::new(MetricsRegistry::new())),
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let resp = readiness(State(state)).await;
-        assert_eq!(resp.status(), StatusCode::OK);
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_pool_closed() {
-        let engine = fresh_engine();
-        engine.pool.close();
-        let state = Arc::new(AppState {
-            engine: engine_swap(engine),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let body = Bytes::from(vec![0u8; 100]);
-        let result = transcribe(State(state), Query(ExportParams::default()), body).await;
-        match result {
-            Err(resp) => assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE),
-            Ok(_) => panic!("expected pool_closed error"),
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_stream_invalid_audio() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let body = Bytes::from(vec![0u8; 100]);
-        let result = transcribe_stream(State(state), body).await;
-        match result {
-            Err(resp) => assert_eq!(resp.status(), StatusCode::UNPROCESSABLE_ENTITY),
-            Ok(_) => panic!("expected invalid_audio error"),
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_stream_payload_too_large() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits {
-                body_limit_bytes: 10,
-                ..RuntimeLimits::default()
-            })),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let body = Bytes::from(vec![0u8; 100]);
-        let result = transcribe_stream(State(state), body).await;
-        match result {
-            Err(resp) => assert_eq!(resp.status(), StatusCode::PAYLOAD_TOO_LARGE),
-            Ok(_) => panic!("expected payload_too_large error"),
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_stream_pool_closed() {
-        let engine = fresh_engine();
-        engine.pool.close();
-        let state = Arc::new(AppState {
-            engine: engine_swap(engine),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let body = minimal_wav();
-        let result = transcribe_stream(State(state), body).await;
-        match result {
-            Err(resp) => assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE),
-            Ok(_) => panic!("expected pool_closed error"),
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_with_metrics() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: Some(Arc::new(MetricsRegistry::new())),
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let body = short_wav();
-        match transcribe(State(state), Query(ExportParams::default()), body).await {
-            Ok(_) => {}
-            Err(_) => panic!("transcribe with metrics failed"),
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_stream_with_metrics() {
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: Some(Arc::new(MetricsRegistry::new())),
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let body = short_wav();
-        match transcribe_stream(State(state), body).await {
-            Ok(_) => {}
-            Err(_) => panic!("transcribe_stream with metrics failed"),
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_transcribe_segments_json() {
-        // `?segments=true` on the default JSON response adds a `segments` array
-        // with sane start/end ordering and per-segment words, while keeping the
-        // top-level text/words/duration contract.
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let params = ExportParams {
-            segments: Some(true),
-            ..ExportParams::default()
-        };
-        let resp = transcribe(State(state), Query(params), short_wav())
-            .await
-            .expect("transcribe with segments should succeed");
-        assert_eq!(resp.status(), StatusCode::OK);
-        let bytes = axum::body::to_bytes(resp.into_body(), 1 << 20)
-            .await
-            .unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        // Top-level contract is preserved.
-        assert!(v.get("text").is_some());
-        assert!(v.get("words").is_some());
-        assert!(v.get("duration").is_some());
-        // The segments array is present and every segment has monotonic timing.
-        let segments = v["segments"].as_array().expect("segments array present");
-        for seg in segments {
-            let start = seg["start"].as_f64().unwrap();
-            let end = seg["end"].as_f64().unwrap();
-            assert!(end >= start, "segment end {end} < start {start}");
-            assert!(seg["words"].is_array());
-        }
-    }
-
-    fn sample_export_result() -> gigastt_core::inference::TranscribeResult {
-        use gigastt_core::inference::WordInfo;
-        gigastt_core::inference::TranscribeResult {
-            text: "привет мир".into(),
-            words: vec![
-                WordInfo::new("привет", 0.0, 0.5, 0.98, Some(0)),
-                WordInfo::new("мир", 0.6, 1.0, 0.97, Some(0)),
-            ],
-            duration_s: 1.0,
-            confidence: None,
-        }
-    }
-
-    #[tokio::test]
-    async fn test_render_export_default_returns_none() {
-        let result = sample_export_result();
-        let params = ExportParams::default();
-        assert!(render_export_response(&result, &params).unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn test_render_export_txt() {
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("txt".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        assert_eq!(body, "привет мир");
-    }
-
-    #[tokio::test]
-    async fn test_render_export_srt_content_type() {
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("srt".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers().get(header::CONTENT_TYPE).unwrap(),
-            "application/x-subrip; charset=utf-8"
-        );
-        let body = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        let text = String::from_utf8(body.to_vec()).unwrap();
-        assert!(text.contains("[SPEAKER_0] привет мир"));
-    }
-
-    #[tokio::test]
-    async fn test_render_export_vtt_download_header() {
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("vtt".into()),
-            download: Some("recording.vtt".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(
-            resp.headers().get(header::CONTENT_DISPOSITION).unwrap(),
-            "attachment; filename=\"recording.vtt\"; filename*=UTF-8''recording.vtt"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_export_download_filename_with_control_char_does_not_panic() {
-        // The download filename is user-controlled; control characters must not
-        // produce an invalid header value / panic — they are sanitized out of
-        // the quoted fallback and percent-encoded in `filename*`.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("srt".into()),
-            download: Some("evil\r\nInjected: x".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers().get(header::CONTENT_DISPOSITION).unwrap(),
-            "attachment; filename=\"evil__Injected: x\"; filename*=UTF-8''evil%0D%0AInjected%3A%20x"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_export_invalid_format() {
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("docx".into()),
-            ..ExportParams::default()
-        };
-        let err = render_export_response(&result, &params).unwrap_err();
-        assert_eq!(err.status(), StatusCode::BAD_REQUEST);
-    }
-
-    #[tokio::test]
-    async fn test_render_export_invalid_format_body_code() {
-        // The invalid-format error carries the machine-readable `invalid_format`
-        // code so clients can distinguish it from other 400s.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("xml".into()),
-            ..ExportParams::default()
-        };
-        let err = render_export_response(&result, &params).unwrap_err();
-        assert_eq!(err.status(), StatusCode::BAD_REQUEST);
-        let bytes = axum::body::to_bytes(err.into_body(), 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "invalid_format");
-    }
-
-    #[tokio::test]
-    async fn test_render_export_uppercase_json_returns_none() {
-        // Format negotiation is case-insensitive: an explicit (any-case) "json"
-        // means "keep the default TranscribeResponse contract", so the helper
-        // returns None instead of building a Response.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("JSON".into()),
-            ..ExportParams::default()
-        };
-        assert!(render_export_response(&result, &params).unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn test_render_export_uppercase_format_renders() {
-        // Non-JSON format strings are also case-insensitive (parsed via
-        // ExportFormat::from_str), so "SRT" still renders subtitles.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("SRT".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers().get(header::CONTENT_TYPE).unwrap(),
-            "application/x-subrip; charset=utf-8"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_export_empty_download_uses_default_name() {
-        // An empty `download` value still requests an attachment; the helper
-        // synthesizes the default `transcript.<ext>` filename.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("vtt".into()),
-            download: Some(String::new()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(
-            resp.headers().get(header::CONTENT_DISPOSITION).unwrap(),
-            "attachment; filename=\"transcript.vtt\"; filename*=UTF-8''transcript.vtt"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_export_download_filename_injection_neutralized() {
-        // A crafted `download` value trying to splice a second `filename*`
-        // parameter must survive only as inert data: the quote becomes `_` in
-        // the fallback, and the `filename*` bytes are percent-encoded, so the
-        // spoofed `spoofed.exe` never appears as a real header parameter.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("srt".into()),
-            download: Some("evil\"; filename*=UTF-8''spoofed.exe".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(
-            resp.headers().get(header::CONTENT_DISPOSITION).unwrap(),
-            "attachment; filename=\"evil_; filename*=UTF-8''spoofed.exe\"; \
-             filename*=UTF-8''evil%22%3B%20filename%2A%3DUTF-8%27%27spoofed.exe"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_export_download_filename_unicode_percent_encoded() {
-        // Non-ASCII names get an ASCII-safe fallback for legacy clients and
-        // keep the full UTF-8 name percent-encoded in `filename*` (RFC 6266).
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("txt".into()),
-            download: Some("é.txt".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(
-            resp.headers().get(header::CONTENT_DISPOSITION).unwrap(),
-            "attachment; filename=\"_.txt\"; filename*=UTF-8''%C3%A9.txt"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_export_download_filename_cyrillic_percent_encoded() {
-        // Cyrillic input must not leak raw non-ASCII bytes into the header:
-        // the fallback replaces each non-ASCII character, `filename*` carries
-        // the percent-encoded UTF-8, and the whole value stays ASCII.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("srt".into()),
-            download: Some("отчёт.srt".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        let value = resp
-            .headers()
-            .get(header::CONTENT_DISPOSITION)
-            .unwrap()
-            .to_str()
-            .unwrap();
-        let encoded_name = format!("{}.srt", "%d0%be%d1%82%d1%87%d1%91%d1%82".to_uppercase());
-        assert_eq!(
-            value,
-            format!("attachment; filename=\"_____.srt\"; filename*=UTF-8''{encoded_name}")
-        );
-        assert!(value.is_ascii());
-    }
-
-    #[tokio::test]
-    async fn test_render_export_md_includes_word_timestamps() {
-        // The Markdown path honours `word_timestamps` and renders the per-word
-        // table; the content type is text/markdown.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("md".into()),
-            word_timestamps: Some(true),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers().get(header::CONTENT_TYPE).unwrap(),
-            "text/markdown; charset=utf-8"
-        );
-        let body = axum::body::to_bytes(resp.into_body(), 4096).await.unwrap();
-        let text = String::from_utf8(body.to_vec()).unwrap();
-        assert!(text.contains("# Transcript"));
-        assert!(text.contains("| Word | Start | End |"));
-    }
-
-    #[tokio::test]
-    async fn test_render_export_line_break_opts_passed_through() {
-        // Tight per-line caps must be threaded into RenderOpts so the rendered
-        // subtitles actually break — proving the params override the defaults.
-        let result = sample_export_result();
-        let loose = ExportParams {
-            format: Some("srt".into()),
-            ..ExportParams::default()
-        };
-        let tight = ExportParams {
-            format: Some("srt".into()),
-            max_words_per_line: Some(1),
-            ..ExportParams::default()
-        };
-        let loose_resp = render_export_response(&result, &loose).unwrap().unwrap();
-        let tight_resp = render_export_response(&result, &tight).unwrap().unwrap();
-        let loose_body = axum::body::to_bytes(loose_resp.into_body(), 4096)
-            .await
-            .unwrap();
-        let tight_body = axum::body::to_bytes(tight_resp.into_body(), 4096)
-            .await
-            .unwrap();
-        let loose_text = String::from_utf8(loose_body.to_vec()).unwrap();
-        let tight_text = String::from_utf8(tight_body.to_vec()).unwrap();
-        // One word per line yields one cue per word (more "-->" arrows) than the
-        // default 14-words-per-line grouping.
-        let loose_cues = loose_text.matches("-->").count();
-        let tight_cues = tight_text.matches("-->").count();
-        assert!(
-            tight_cues > loose_cues,
-            "tight={tight_cues} should exceed loose={loose_cues}"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_render_export_md_segments_emits_headers() {
-        // `format=md` + `segments=true` switches Markdown to `### [mm:ss]`
-        // section headers over the cue boundaries, dropping the flat
-        // `# Transcript` blob.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("md".into()),
-            segments: Some(true),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers().get(header::CONTENT_TYPE).unwrap(),
-            "text/markdown; charset=utf-8"
-        );
-        let body = axum::body::to_bytes(resp.into_body(), 4096).await.unwrap();
-        let text = String::from_utf8(body.to_vec()).unwrap();
-        assert!(text.contains("### [00:00]"));
-        assert!(text.contains("[SPEAKER_0] привет мир"));
-        // Segment mode replaces the flat transcript blob.
-        assert!(!text.contains("# Transcript"));
-    }
-
-    #[tokio::test]
-    async fn test_render_export_md_without_segments_unchanged() {
-        // Plain `format=md` (no segments) keeps the existing frontmatter +
-        // `# Transcript` layout — segment mode is strictly opt-in.
-        let result = sample_export_result();
-        let params = ExportParams {
-            format: Some("md".into()),
-            ..ExportParams::default()
-        };
-        let resp = render_export_response(&result, &params).unwrap().unwrap();
-        let body = axum::body::to_bytes(resp.into_body(), 4096).await.unwrap();
-        let text = String::from_utf8(body.to_vec()).unwrap();
-        assert!(text.contains("# Transcript"));
-        assert!(!text.contains("### ["));
-    }
-
-    #[tokio::test]
-    async fn test_render_export_segments_ignored_for_srt() {
-        // `segments=true` is a JSON/Markdown affordance; SRT is already
-        // cue-based and must render identically with or without the flag.
-        let result = sample_export_result();
-        let plain = ExportParams {
-            format: Some("srt".into()),
-            ..ExportParams::default()
-        };
-        let with_segments = ExportParams {
-            format: Some("srt".into()),
-            segments: Some(true),
-            ..ExportParams::default()
-        };
-        let a = render_export_response(&result, &plain).unwrap().unwrap();
-        let b = render_export_response(&result, &with_segments)
-            .unwrap()
-            .unwrap();
-        let a_body = axum::body::to_bytes(a.into_body(), 4096).await.unwrap();
-        let b_body = axum::body::to_bytes(b.into_body(), 4096).await.unwrap();
-        assert_eq!(a_body, b_body);
-    }
-
-    #[test]
-    fn test_transcribe_response_omits_segments_when_none() {
-        // The default response must be byte-identical to the pre-feature shape:
-        // no `segments` key when the caller didn't ask for it.
-        let resp = TranscribeResponse {
-            text: "hello".into(),
-            words: vec![],
-            duration: 1.5,
-            confidence: None,
-            segments: None,
-        };
-        let v = serde_json::to_value(&resp).unwrap();
-        assert!(v.get("segments").is_none());
-        assert_eq!(v["text"], "hello");
-        assert_eq!(v["duration"], 1.5);
-    }
-
-    #[test]
-    fn test_transcribe_response_confidence_present_only_when_some() {
-        // With words decoded, the aggregate rides the top-level response;
-        // without words the key is omitted so the response shape matches the
-        // pre-field contract exactly.
-        let with = TranscribeResponse {
-            text: "hello".into(),
-            words: vec![],
-            duration: 1.5,
-            confidence: Some(0.87),
-            segments: None,
-        };
-        let v = serde_json::to_value(&with).unwrap();
-        let c = v["confidence"].as_f64().expect("numeric confidence");
-        assert!((c - 0.87).abs() < 1e-6, "got {c}");
-
-        let without = TranscribeResponse {
-            text: String::new(),
-            words: vec![],
-            duration: 0.0,
-            confidence: None,
-            segments: None,
-        };
-        let v = serde_json::to_value(&without).unwrap();
-        assert!(v.get("confidence").is_none());
-    }
-
-    #[test]
-    fn test_transcribe_response_includes_segments_when_present() {
-        use gigastt_core::export::to_segments;
-        use gigastt_core::inference::WordInfo;
-        let words = vec![
-            WordInfo::new("привет", 0.0, 0.5, 0.98, None),
-            WordInfo::new("мир", 0.6, 1.0, 0.97, None),
-        ];
-        let resp = TranscribeResponse {
-            text: "привет мир".into(),
-            words: words.clone(),
-            duration: 1.0,
-            confidence: None,
-            segments: Some(to_segments(&words, 80, 14)),
-        };
-        let v = serde_json::to_value(&resp).unwrap();
-        let segments = v["segments"].as_array().unwrap();
-        assert_eq!(segments.len(), 1);
-        assert_eq!(segments[0]["start"], 0.0);
-        assert_eq!(segments[0]["end"], 1.0);
-        assert_eq!(segments[0]["text"], "привет мир");
-        assert_eq!(segments[0]["words"][0]["word"], "привет");
-    }
-
-    #[test]
-    fn test_export_params_deserialize_from_query() {
-        // The query-param shape drives format negotiation; confirm axum's Query
-        // extractor maps every field so the handler sees the caller's choices.
-        let uri: axum::http::Uri = "http://x/?format=srt&download=out.srt&max_chars_per_line=20&max_words_per_line=3&word_timestamps=true&segments=true&channels=split&diarization=true"
-            .parse()
-            .unwrap();
-        let Query(params): Query<ExportParams> = Query::try_from_uri(&uri).unwrap();
-        assert_eq!(params.format.as_deref(), Some("srt"));
-        assert_eq!(params.download.as_deref(), Some("out.srt"));
-        assert_eq!(params.max_chars_per_line, Some(20));
-        assert_eq!(params.max_words_per_line, Some(3));
-        assert_eq!(params.word_timestamps, Some(true));
-        assert_eq!(params.segments, Some(true));
-        assert_eq!(params.channels.as_deref(), Some("split"));
-        assert_eq!(params.diarization, Some(true));
-    }
-
-    #[test]
-    fn test_export_params_default_empty_query() {
-        // No query params -> all None, which the handler maps to JSON defaults.
-        let uri: axum::http::Uri = "http://x/".parse().unwrap();
-        let Query(params): Query<ExportParams> = Query::try_from_uri(&uri).unwrap();
-        assert!(params.format.is_none());
-        assert!(params.download.is_none());
-        assert!(params.max_chars_per_line.is_none());
-        // The per-request knob overrides default to absent so the handler falls
-        // back to the engine's boot policy (byte-unchanged response).
-        assert!(params.punctuation.is_none());
-        assert!(params.itn.is_none());
-        assert!(params.vad.is_none());
-        assert!(params.hotwords.is_none());
-        assert!(params.hotwords_boost.is_none());
-        assert!(params.variant.is_none());
-    }
-
-    #[test]
-    fn test_transcribe_knob_params_deserialize_from_query() {
-        // `?punctuation=false&itn=false&vad=false&variant=rnnt` maps to
-        // `Some(false)`/`Some("rnnt")`, letting the handler override the boot
-        // policy per request.
-        let uri: axum::http::Uri = "http://x/?punctuation=false&itn=false&vad=false&variant=rnnt"
-            .parse()
-            .unwrap();
-        let Query(params): Query<ExportParams> = Query::try_from_uri(&uri).unwrap();
-        assert_eq!(params.punctuation, Some(false));
-        assert_eq!(params.itn, Some(false));
-        assert_eq!(params.vad, Some(false));
-        assert_eq!(params.variant.as_deref(), Some("rnnt"));
-
-        // The `true` direction deserializes symmetrically.
-        let uri: axum::http::Uri = "http://x/?punctuation=true&itn=true&vad=true"
-            .parse()
-            .unwrap();
-        let Query(params): Query<ExportParams> = Query::try_from_uri(&uri).unwrap();
-        assert_eq!(params.punctuation, Some(true));
-        assert_eq!(params.itn, Some(true));
-        assert_eq!(params.vad, Some(true));
-    }
-
-    #[test]
-    fn test_hotwords_query_param_parsing() {
-        // Comma-separated phrases + optional boost deserialize from the query
-        // string and map to HotwordOverride via hotwords_from_export_params.
-        let uri: axum::http::Uri =
-            "http://x/?hotwords=%D1%81%D0%B1%D0%B5%D1%80,%D1%82%D0%B8%D0%BD%D1%8C%D0%BA%D0%BE%D1%84%D1%84&hotwords_boost=3.5"
-                .parse()
-                .unwrap();
-        let Query(params): Query<ExportParams> = Query::try_from_uri(&uri).unwrap();
-        assert_eq!(params.hotwords.as_deref(), Some("сбер,тинькофф"));
-        assert_eq!(params.hotwords_boost, Some(3.5));
-
-        let hw = hotwords_from_export_params(&params).expect("hotwords present");
-        assert_eq!(hw.phrases, vec!["сбер".to_string(), "тинькофф".to_string()]);
-        assert_eq!(hw.boost, Some(3.5));
-
-        // Absent hotwords → engine default (None), even if boost is set alone.
-        let uri: axum::http::Uri = "http://x/?hotwords_boost=9".parse().unwrap();
-        let Query(params): Query<ExportParams> = Query::try_from_uri(&uri).unwrap();
-        assert!(hotwords_from_export_params(&params).is_none());
-
-        // Empty key present → Some(empty) force-off.
-        let uri: axum::http::Uri = "http://x/?hotwords=".parse().unwrap();
-        let Query(params): Query<ExportParams> = Query::try_from_uri(&uri).unwrap();
-        let hw = hotwords_from_export_params(&params).expect("key present means override");
-        assert!(hw.phrases.is_empty());
-
-        assert_eq!(
-            parse_hotwords_query(" сбер , , тинькофф "),
-            vec!["сбер".to_string(), "тинькофф".to_string()]
-        );
-        assert!(parse_hotwords_query("").is_empty());
-        assert!(parse_hotwords_query(",,,").is_empty());
-    }
-
-    #[test]
-    fn test_model_info_serialization_shape() {
-        // ModelInfo is the /v1/models contract; assert the field names/values
-        // clients depend on are present and correctly typed.
-        let info = ModelInfo {
-            id: "gigaam-v3-rnnt".into(),
-            name: "GigaAM v3 RNN-T".into(),
-            variant: "rnnt".into(),
-            version: "0.9.0".into(),
-            encoder: "int8".into(),
-            vocab_size: 34,
-            sample_rate: 16000,
-            pool_size: 4,
-            pool_available: 3,
-            supported_formats: vec!["wav".into(), "mp3".into()],
-            supported_rates: vec![16000, 48000],
-            punctuation: true,
-            itn: true,
-            diarization: false,
-        };
-        let v = serde_json::to_value(&info).unwrap();
-        assert_eq!(v["id"], "gigaam-v3-rnnt");
-        assert_eq!(v["variant"], "rnnt");
-        assert_eq!(v["encoder"], "int8");
-        assert_eq!(v["vocab_size"], 34);
-        assert_eq!(v["sample_rate"], 16000);
-        assert_eq!(v["punctuation"], true);
-        assert_eq!(v["itn"], true);
-        assert_eq!(v["diarization"], false);
-        assert_eq!(v["supported_rates"][1], 48000);
-    }
-
-    #[tokio::test]
-    async fn test_api_inference_timeout_error_body_message() {
-        // The 504 inference-timeout body should not leak internals, just the
-        // stable code + a sanitized message.
-        let resp = api_inference_timeout_error();
-        let bytes = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "inference_timeout");
-        assert_eq!(v["error"], "Inference timed out.");
-    }
-
-    #[tokio::test]
-    async fn test_api_pool_closed_error_status_and_message() {
-        // pool_closed is a 503 with a sanitized "shutting down" message and no
-        // retry hint (the pool is not coming back).
-        let resp = api_pool_closed_error();
-        let (parts, body) = resp.into_parts();
-        assert_eq!(parts.status, StatusCode::SERVICE_UNAVAILABLE);
-        let bytes = axum::body::to_bytes(body, 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["error"], "Server is shutting down");
-        assert_eq!(v["code"], "pool_closed");
-    }
-
-    #[test]
-    fn test_peer_is_loopback_guard() {
-        use std::net::{Ipv4Addr, Ipv6Addr, SocketAddr};
-        // IPv4 + IPv6 loopback are accepted regardless of source port.
-        assert!(peer_is_loopback(&SocketAddr::from((
-            Ipv4Addr::LOCALHOST,
-            5000
-        ))));
-        assert!(peer_is_loopback(&SocketAddr::from((
-            Ipv6Addr::LOCALHOST,
-            5000
-        ))));
-        // A non-loopback peer (LAN / public) is rejected — reload must stay local
-        // even under --bind-all / --cors-allow-any.
-        assert!(!peer_is_loopback(&SocketAddr::from((
-            Ipv4Addr::new(192, 168, 1, 10),
-            9876
-        ))));
-        assert!(!peer_is_loopback(&SocketAddr::from((
-            Ipv4Addr::new(10, 0, 0, 1),
-            9876
-        ))));
-        assert!(!peer_is_loopback(&SocketAddr::from((
-            Ipv4Addr::new(8, 8, 8, 8),
-            443
-        ))));
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_reload_rejects_non_loopback_peer() {
-        // The loopback guard fires before any engine work: a non-loopback
-        // ConnectInfo yields 403 `loopback_only` even with a builder present.
-        // Model-gated only because `AppState` needs a concrete `Engine`; the
-        // pure guard logic is covered model-free by `test_peer_is_loopback_guard`.
-        use std::net::{Ipv4Addr, SocketAddr};
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            engine_builder: Some(Arc::new(|| {
-                anyhow::bail!("builder must not run for a rejected peer")
-            })),
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let peer = SocketAddr::from((Ipv4Addr::new(203, 0, 113, 7), 40000));
-        let resp = reload(axum::extract::ConnectInfo(peer), State(state)).await;
-        assert_eq!(resp.status(), StatusCode::FORBIDDEN);
-        let bytes = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "loopback_only");
-    }
-
-    #[tokio::test]
-    #[ignore = "requires model"]
-    async fn test_reload_unsupported_when_no_builder() {
-        // A loopback peer with no stored builder (the thin `run_with_shutdown` /
-        // test path) gets `reload_unsupported`, not a swap.
-        use std::net::{Ipv4Addr, SocketAddr};
-        let state = Arc::new(AppState {
-            engine: engine_swap(test_engine()),
-            engine_builder: None,
-            reload_lock: Arc::new(tokio::sync::Mutex::new(())),
-            limits: Arc::new(ArcSwap::from_pointee(RuntimeLimits::default())),
-            metrics_registry: None,
-            shutdown: tokio_util::sync::CancellationToken::new(),
-            tracker: tokio_util::task::TaskTracker::new(),
-            jobs: None,
-        });
-        let peer = SocketAddr::from((Ipv4Addr::LOCALHOST, 40000));
-        let resp = reload(axum::extract::ConnectInfo(peer), State(state)).await;
-        assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE);
-        let bytes = axum::body::to_bytes(resp.into_body(), 1024).await.unwrap();
-        let v: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-        assert_eq!(v["code"], "reload_unsupported");
-    }
-
-    #[test]
-    fn test_sse_data_payload_includes_words_and_timestamp() {
-        // A successful segment carries text, timestamp and words through
-        // unchanged so SSE clients can render word-level UI.
-        use gigastt_core::inference::WordInfo;
-        let mut seg = gigastt_core::inference::TranscriptSegment::empty_final();
-        seg.text = "привет".into();
-        seg.timestamp = 1.25;
-        seg.words = vec![WordInfo::new("привет", 0.0, 0.5, 0.99, Some(0))];
-        let payload = sse_data_payload(&Ok(seg));
-        let v: serde_json::Value = serde_json::from_str(&payload).unwrap();
-        assert_eq!(v["type"], "final");
-        assert_eq!(v["text"], "привет");
-        assert_eq!(v["timestamp"], 1.25);
-        assert_eq!(v["words"][0]["word"], "привет");
-    }
-
-    fn test_engine() -> Arc<Engine> {
-        use std::sync::OnceLock;
-        static ENGINE: OnceLock<Arc<Engine>> = OnceLock::new();
-        ENGINE
-            .get_or_init(|| {
-                Arc::new(
-                    Engine::load_with_pool_size(&gigastt_core::model::default_model_dir(), 1)
-                        .unwrap(),
-                )
-            })
-            .clone()
-    }
-
-    fn fresh_engine() -> Arc<Engine> {
-        Arc::new(Engine::load_with_pool_size(&gigastt_core::model::default_model_dir(), 1).unwrap())
-    }
-
-    /// Wrap an engine handle in the `ArcSwap` the `AppState` now holds. Keeps
-    /// the model-gated test constructors terse after the hot-reload swap change.
-    fn engine_swap(engine: Arc<Engine>) -> Arc<ArcSwap<Engine>> {
-        Arc::new(ArcSwap::new(engine))
-    }
-
-    fn minimal_wav() -> Bytes {
-        let data_size = 4u32;
-        let file_size = 44 + data_size - 8;
-        let mut wav = vec![];
-        wav.extend_from_slice(b"RIFF");
-        wav.extend_from_slice(&file_size.to_le_bytes());
-        wav.extend_from_slice(b"WAVE");
-        wav.extend_from_slice(b"fmt ");
-        wav.extend_from_slice(&16u32.to_le_bytes());
-        wav.extend_from_slice(&1u16.to_le_bytes());
-        wav.extend_from_slice(&1u16.to_le_bytes());
-        wav.extend_from_slice(&16000u32.to_le_bytes());
-        wav.extend_from_slice(&(16000u32 * 2).to_le_bytes());
-        wav.extend_from_slice(&2u16.to_le_bytes());
-        wav.extend_from_slice(&16u16.to_le_bytes());
-        wav.extend_from_slice(b"data");
-        wav.extend_from_slice(&data_size.to_le_bytes());
-        wav.extend_from_slice(&0i16.to_le_bytes());
-        wav.extend_from_slice(&0i16.to_le_bytes());
-        Bytes::from(wav)
-    }
-
-    fn short_wav() -> Bytes {
-        let sample_rate = 16000u32;
-        let duration_s = 0.1f32;
-        let num_samples = (sample_rate as f32 * duration_s) as u32;
-        let data_size = num_samples * 2;
-        let file_size = 44 + data_size - 8;
-        let mut wav = vec![];
-        wav.extend_from_slice(b"RIFF");
-        wav.extend_from_slice(&file_size.to_le_bytes());
-        wav.extend_from_slice(b"WAVE");
-        wav.extend_from_slice(b"fmt ");
-        wav.extend_from_slice(&16u32.to_le_bytes());
-        wav.extend_from_slice(&1u16.to_le_bytes());
-        wav.extend_from_slice(&1u16.to_le_bytes());
-        wav.extend_from_slice(&sample_rate.to_le_bytes());
-        wav.extend_from_slice(&(sample_rate * 2).to_le_bytes());
-        wav.extend_from_slice(&2u16.to_le_bytes());
-        wav.extend_from_slice(&16u16.to_le_bytes());
-        wav.extend_from_slice(b"data");
-        wav.extend_from_slice(&data_size.to_le_bytes());
-        for _ in 0..num_samples {
-            wav.extend_from_slice(&0i16.to_le_bytes());
-        }
-        Bytes::from(wav)
-    }
-}
